https://telecominfraproject.atlassian.net/browse/WIFI-11280
