+++ src/Daemon.cpp
@@ -28,6 +28,7 @@
 #include "VenueBroadcaster.h"
 #include "framework/ConfigurationValidator.h"
 #include "rttys/RTTYS_server.h"
+#include "framework/UI_WebSocketClientServer.h"
 
 namespace OpenWifi {
 	class Daemon *Daemon::instance() {
@@ -40,10 +41,9 @@ namespace OpenWifi {
 										StorageService(),
 										SerialNumberCache(),
 										ConfigurationValidator(),
-								   		WebSocketClientServer(),
+								   		UI_WebSocketClientServer(),
 										OUIServer(),
 										FindCountryFromIP(),
-//										DeviceRegistry(),
 										CommandManager(),
 										FileUploader(),
 										StorageArchiver(),
