+++ src/framework/MicroService.h
@@ -99,19 +99,27 @@ using namespace std::chrono_literals;
 #include "Poco/ThreadLocal.h"
 #include "Poco/NullChannel.h"
 
-#include "cppkafka/cppkafka.h"
-
 #include "framework/MicroServiceErrorHandler.h"
 #include "framework/OpenWifiTypes.h"
 #include "framework/KafkaTopics.h"
 #include "framework/ow_constants.h"
+#include "framework/AppServiceRegistry.h"
+#include "framework/utils.h"
+#include "framework/SubSystemServer.h"
+#include "framework/UI_WebSocketClientServer.h"
+#include "framework/MicroServiceNames.h"
+#include "framework/AuthClient.h"
+
 #include "RESTObjects/RESTAPI_SecurityObjects.h"
+#include "cppkafka/cppkafka.h"
+
 #include "nlohmann/json.hpp"
 #include "ow_version.h"
 #include "fmt/core.h"
 
 #define _OWDEBUG_ std::cout<< __FILE__ <<":" << __LINE__ << std::endl;
 // #define _OWDEBUG_ Logger().debug(Poco::format("%s: %lu",__FILE__,__LINE__));
+
 namespace OpenWifi {
 
     enum UNAUTHORIZED_REASON {
@@ -131,1118 +139,10 @@ namespace OpenWifi {
         SECURITY_SERVICE_UNREACHABLE,
         CANNOT_REFRESH_TOKEN
     };
-
-	class AppServiceRegistry {
-	  public:
-		inline AppServiceRegistry();
-
-		static AppServiceRegistry & instance() {
-		    static auto instance_= new AppServiceRegistry;
-			return *instance_;
-		}
-
-		inline ~AppServiceRegistry() {
-			Save();
-		}
-
-		inline void Save() {
-			std::istringstream  IS( to_string(Registry_));
-			std::ofstream       OF;
-			OF.open(FileName,std::ios::binary | std::ios::trunc);
-			Poco::StreamCopier::copyStream(IS, OF);
-		}
-
-		inline void Set(const char *Key, uint64_t Value ) {
-			Registry_[Key] = Value;
-			Save();
-		}
-
-		inline void Set(const char *Key, const std::string &Value ) {
-			Registry_[Key] = Value;
-			Save();
-		}
-
-		inline void Set(const char *Key, bool Value ) {
-			Registry_[Key] = Value;
-			Save();
-		}
-
-		inline bool Get(const char *Key, bool & Value ) {
-			if(Registry_[Key].is_boolean()) {
-				Value = Registry_[Key].get<bool>();
-				return true;
-			}
-			return false;
-		}
-
-		inline bool Get(const char *Key, uint64_t & Value ) {
-			if(Registry_[Key].is_number_unsigned()) {
-				Value = Registry_[Key].get<uint64_t>();
-				return true;
-			}
-			return false;
-		}
-
-		inline bool Get(const char *Key, std::string & Value ) {
-			if(Registry_[Key].is_string()) {
-				Value = Registry_[Key].get<std::string>();
-				return true;
-			}
-			return false;
-		}
-
-	  private:
-		std::string         FileName;
-		nlohmann::json      Registry_;
-	};
-
-    inline auto AppServiceRegistry() { return AppServiceRegistry::instance(); }
-
 }
 
-namespace OpenWifi::RESTAPI_utils {
-
-    inline void EmbedDocument(const std::string & ObjName, Poco::JSON::Object & Obj, const std::string &ObjStr) {
-        std::string D = ObjStr.empty() ? "{}" : ObjStr;
-        Poco::JSON::Parser P;
-        Poco::Dynamic::Var result = P.parse(D);
-        const auto &DetailsObj = result.extract<Poco::JSON::Object::Ptr>();
-        Obj.set(ObjName, DetailsObj);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, bool V) {
-        Obj.set(Field,V);
-    }
-
-	inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, double V) {
-		Obj.set(Field,V);
-	}
-
-	inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, float V) {
-		Obj.set(Field,V);
-	}
-
-	inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, const std::string & S) {
-        Obj.set(Field,S);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, const char * S) {
-        Obj.set(Field,S);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, int16_t Value) {
-        Obj.set(Field, Value);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, int32_t Value) {
-        Obj.set(Field, Value);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, int64_t Value) {
-        Obj.set(Field, Value);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, uint16_t Value) {
-        Obj.set(Field, Value);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, uint32_t Value) {
-        Obj.set(Field, Value);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, uint64_t Value) {
-        Obj.set(Field,Value);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, const Poco::Data::BLOB &Value) {
-        auto Result = Utils::base64encode((const unsigned char *)Value.rawContent(),Value.size());
-        Obj.set(Field,Result);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, const Types::StringPairVec & S) {
-        Poco::JSON::Array   Array;
-        for(const auto &i:S) {
-            Poco::JSON::Object  O;
-            O.set("tag",i.first);
-            O.set("value", i.second);
-            Array.add(O);
-        }
-        Obj.set(Field,Array);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, const Types::StringVec &V) {
-        Poco::JSON::Array	A;
-        for(const auto &i:V)
-            A.add(i);
-        Obj.set(Field,A);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, const Types::TagList &V) {
-        Poco::JSON::Array	A;
-        for(const auto &i:V)
-            A.add(i);
-        Obj.set(Field,A);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, const Types::CountedMap &M) {
-        Poco::JSON::Array	A;
-        for(const auto &[Key,Value]:M) {
-            Poco::JSON::Object  O;
-            O.set("tag",Key);
-            O.set("value", Value);
-            A.add(O);
-        }
-        Obj.set(Field,A);
-    }
-
-    inline void field_to_json(Poco::JSON::Object &Obj, const char *Field, const Types::Counted3DMapSII &M) {
-        Poco::JSON::Array	A;
-        for(const auto &[OrgName,MonthlyNumberMap]:M) {
-            Poco::JSON::Object  OrgObject;
-            OrgObject.set("tag",OrgName);
-            Poco::JSON::Array   MonthlyArray;
-            for(const auto &[Month,Counter]:MonthlyNumberMap) {
-                Poco::JSON::Object  Inner;
-                Inner.set("value", Month);
-                Inner.set("counter", Counter);
-                MonthlyArray.add(Inner);
-            }
-            OrgObject.set("index",MonthlyArray);
-            A.add(OrgObject);
-        }
-        Obj.set(Field, A);
-    }
-
-    template<typename T> void field_to_json(Poco::JSON::Object &Obj,
-            const char *Field,
-            const T &V,
-            std::function<std::string(const T &)> F) {
-        Obj.set(Field, F(V));
-    }
-
-    template<class T> void field_to_json(Poco::JSON::Object &Obj, const char *Field, const std::vector<T> &Value) {
-        Poco::JSON::Array Arr;
-        for(const auto &i:Value) {
-            Poco::JSON::Object	AO;
-            i.to_json(AO);
-            Arr.add(AO);
-        }
-        Obj.set(Field, Arr);
-    }
-
-    template<class T> void field_to_json(Poco::JSON::Object &Obj, const char *Field, const T &Value) {
-        Poco::JSON::Object  Answer;
-        Value.to_json(Answer);
-        Obj.set(Field, Answer);
-    }
-
-    ///////////////////////////
-    ///////////////////////////
-    ///////////////////////////
-    ///////////////////////////
-
-    template<typename T> bool field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, T & V,
-            std::function<T(const std::string &)> F) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            V = F(Obj->get(Field).toString());
-        return true;
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, std::string &S) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            S = Obj->get(Field).toString();
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, double & Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            Value = (double)Obj->get(Field);
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, float & Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            Value = (float)Obj->get(Field);
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, bool &Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            Value = (Obj->get(Field).toString() == "true");
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, int16_t &Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            Value = (int16_t)Obj->get(Field);
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, int32_t &Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            Value = (int32_t) Obj->get(Field);
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, int64_t &Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            Value = (int64_t)Obj->get(Field);
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, uint16_t &Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            Value = (uint16_t)Obj->get(Field);
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, uint32_t &Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            Value = (uint32_t)Obj->get(Field);
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, uint64_t &Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field))
-            Value = (uint64_t)Obj->get(Field);
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, Poco::Data::BLOB &Value) {
-        if(Obj->has(Field) && !Obj->isNull(Field)) {
-            auto Result = Utils::base64decode(Obj->get(Field).toString());
-            Value.assignRaw((const unsigned char *)&Result[0],Result.size());
-        }
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, Types::StringPairVec &Vec) {
-        if(Obj->isArray(Field) && !Obj->isNull(Field)) {
-            auto O = Obj->getArray(Field);
-            for(const auto &i:*O) {
-                std::string S1,S2;
-                auto Inner = i.extract<Poco::JSON::Object::Ptr>();
-                if(Inner->has("tag"))
-                    S1 = Inner->get("tag").toString();
-                if(Inner->has("value"))
-                    S2 = Inner->get("value").toString();
-                auto P = std::make_pair(S1,S2);
-                Vec.push_back(P);
-            }
-        }
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, Types::StringVec &Value) {
-        if(Obj->isArray(Field) && !Obj->isNull(Field)) {
-            Value.clear();
-            Poco::JSON::Array::Ptr A = Obj->getArray(Field);
-            for(const auto &i:*A) {
-                Value.push_back(i.toString());
-            }
-        }
-    }
-
-    inline void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, Types::TagList &Value) {
-        if(Obj->isArray(Field) && !Obj->isNull(Field)) {
-            Value.clear();
-            Poco::JSON::Array::Ptr A = Obj->getArray(Field);
-            for(const auto &i:*A) {
-                Value.push_back(i);
-            }
-        }
-    }
-
-    template<class T> void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, std::vector<T> &Value) {
-        if(Obj->isArray(Field) && !Obj->isNull(Field)) {
-            Poco::JSON::Array::Ptr	Arr = Obj->getArray(Field);
-            for(auto &i:*Arr) {
-                auto InnerObj = i.extract<Poco::JSON::Object::Ptr>();
-                T	NewItem;
-                NewItem.from_json(InnerObj);
-                Value.push_back(NewItem);
-            }
-        }
-    }
-
-    template<class T> void field_from_json(const Poco::JSON::Object::Ptr &Obj, const char *Field, T &Value) {
-        if(Obj->isObject(Field) && !Obj->isNull(Field)) {
-            Poco::JSON::Object::Ptr	A = Obj->getObject(Field);
-            Value.from_json(A);
-        }
-    }
-
-    inline std::string to_string(const Types::TagList & ObjectArray) {
-        Poco::JSON::Array OutputArr;
-        if(ObjectArray.empty())
-            return "[]";
-        for(auto const &i:ObjectArray) {
-            OutputArr.add(i);
-        }
-        std::ostringstream OS;
-        Poco::JSON::Stringifier::stringify(OutputArr,OS, 0,0, Poco::JSON_PRESERVE_KEY_ORDER );
-        return OS.str();
-    }
-
-    inline std::string to_string(const Types::StringVec & ObjectArray) {
-        Poco::JSON::Array OutputArr;
-        if(ObjectArray.empty())
-            return "[]";
-        for(auto const &i:ObjectArray) {
-            OutputArr.add(i);
-        }
-        std::ostringstream OS;
-        Poco::JSON::Stringifier::condense(OutputArr,OS);
-        return OS.str();
-    }
-
-    inline std::string to_string(const Types::StringPairVec & ObjectArray) {
-        Poco::JSON::Array OutputArr;
-        if(ObjectArray.empty())
-            return "[]";
-        for(auto const &i:ObjectArray) {
-            Poco::JSON::Array InnerArray;
-            InnerArray.add(i.first);
-            InnerArray.add(i.second);
-            OutputArr.add(InnerArray);
-        }
-        std::ostringstream OS;
-        Poco::JSON::Stringifier::condense(OutputArr,OS);
-        return OS.str();
-    }
-
-    template<class T> std::string to_string(const std::vector<T> & ObjectArray) {
-        Poco::JSON::Array OutputArr;
-        if(ObjectArray.empty())
-            return "[]";
-        for(auto const &i:ObjectArray) {
-            Poco::JSON::Object O;
-            i.to_json(O);
-            OutputArr.add(O);
-        }
-        std::ostringstream OS;
-        Poco::JSON::Stringifier::condense(OutputArr,OS);
-        return OS.str();
-    }
-
-    template<class T> std::string to_string(const std::vector<std::vector<T>> & ObjectArray) {
-        Poco::JSON::Array OutputArr;
-        if(ObjectArray.empty())
-            return "[]";
-        for(auto const &i:ObjectArray) {
-            Poco::JSON::Array InnerArr;
-            for(auto const &j:i) {
-                if constexpr(std::is_integral<T>::value) {
-                    InnerArr.add(j);
-                } if constexpr(std::is_same_v<T,std::string>) {
-                    InnerArr.add(j);
-                } else {
-                    InnerArr.add(j);
-                    Poco::JSON::Object O;
-                    j.to_json(O);
-                    InnerArr.add(O);
-                }
-            }
-            OutputArr.add(InnerArr);
-        }
-        std::ostringstream OS;
-        Poco::JSON::Stringifier::condense(OutputArr,OS);
-        return OS.str();
-    }
-
-    template<class T> std::string to_string(const T & Object) {
-        Poco::JSON::Object OutputObj;
-        Object.to_json(OutputObj);
-        std::ostringstream OS;
-        Poco::JSON::Stringifier::condense(OutputObj,OS);
-        return OS.str();
-    }
-
-    inline Types::StringVec to_object_array(const std::string & ObjectString) {
-
-        Types::StringVec 	Result;
-        if(ObjectString.empty())
-            return Result;
-
-        try {
-            Poco::JSON::Parser P;
-            auto Object = P.parse(ObjectString).template extract<Poco::JSON::Array::Ptr>();
-            for (auto const &i : *Object) {
-                Result.push_back(i.toString());
-            }
-        } catch (...) {
-
-        }
-        return Result;
-    }
-
-    inline OpenWifi::Types::TagList to_taglist(const std::string & ObjectString) {
-        Types::TagList 	Result;
-        if(ObjectString.empty())
-            return Result;
-
-        try {
-            Poco::JSON::Parser P;
-            auto Object = P.parse(ObjectString).template extract<Poco::JSON::Array::Ptr>();
-            for (auto const &i : *Object) {
-                Result.push_back(i);
-            }
-        } catch (...) {
-
-        }
-        return Result;
-    }
-
-    inline Types::StringPairVec to_stringpair_array(const std::string &S) {
-        Types::StringPairVec   R;
-        if(S.empty())
-            return R;
-        try {
-            Poco::JSON::Parser P;
-            auto Object = P.parse(S).template extract<Poco::JSON::Array::Ptr>();
-            for (const auto &i : *Object) {
-                auto InnerObject = i.template extract<Poco::JSON::Array::Ptr>();
-                if(InnerObject->size()==2) {
-                    auto S1 = InnerObject->getElement<std::string>(0);
-                    auto S2 = InnerObject->getElement<std::string>(1);
-                    R.push_back(std::make_pair(S1,S2));
-                }
-            }
-        } catch (...) {
-
-        }
-
-        return R;
-    }
-
-    template<class T> std::vector<T> to_object_array(const std::string & ObjectString) {
-        std::vector<T>	Result;
-        if(ObjectString.empty())
-            return Result;
-
-        try {
-            Poco::JSON::Parser P;
-            auto Object = P.parse(ObjectString).template extract<Poco::JSON::Array::Ptr>();
-            for (auto const &i : *Object) {
-                auto InnerObject = i.template extract<Poco::JSON::Object::Ptr>();
-                T Obj;
-                Obj.from_json(InnerObject);
-                Result.push_back(Obj);
-            }
-        } catch (...) {
-
-        }
-        return Result;
-    }
-
-    template<class T> std::vector<std::vector<T>> to_array_of_array_of_object(const std::string & ObjectString) {
-        std::vector<std::vector<T>>	Result;
-        if(ObjectString.empty())
-            return Result;
-        try {
-            Poco::JSON::Parser P1;
-            auto OutterArray = P1.parse(ObjectString).template extract<Poco::JSON::Array::Ptr>();
-            for (auto const &i : *OutterArray) {
-                Poco::JSON::Parser P2;
-                auto InnerArray = P2.parse(i).template extract<Poco::JSON::Array::Ptr>();
-                std::vector<T>  InnerVector;
-                for(auto const &j: *InnerArray) {
-                    auto Object = j.template extract<Poco::JSON::Object::Ptr>();
-                    T Obj;
-                    Obj.from_json(Object);
-                    InnerVector.push_back(Obj);
-                }
-                Result.push_back(InnerVector);
-            }
-        } catch (...) {
-
-        }
-        return Result;
-    }
-
-    template<class T> T to_object(const std::string & ObjectString) {
-        T	Result;
-
-        if(ObjectString.empty())
-            return Result;
-
-        Poco::JSON::Parser	P;
-        auto Object = P.parse(ObjectString).template extract<Poco::JSON::Object::Ptr>();
-        Result.from_json(Object);
-
-        return Result;
-    }
-
-    template<class T> bool from_request(T & Obj, Poco::Net::HTTPServerRequest &Request) {
-        Poco::JSON::Parser IncomingParser;
-        auto RawObject = IncomingParser.parse(Request.stream()).extract<Poco::JSON::Object::Ptr>();
-        Obj.from_json(RawObject);
-        return true;
-    }
-}
-
-namespace OpenWifi::Utils {
-
-	inline bool NormalizeMac(std::string & Mac) {
-		Poco::replaceInPlace(Mac,":","");
-		Poco::replaceInPlace(Mac,"-","");
-		if(Mac.size()!=12)
-			return false;
-		for(const auto &i:Mac) {
-			if(!std::isxdigit(i))
-				return false;
-		}
-		Poco::toLowerInPlace(Mac);
-		return true;
-	}
-
-	inline void SetThreadName(const char *name) {
-#ifdef __linux__
-		Poco::Thread::current()->setName(name);
-		pthread_setname_np(pthread_self(), name);
-#endif
-#ifdef __APPLE__
-	Poco::Thread::current()->setName(name);
-	pthread_setname_np(name);
-#endif
-	}
-
-	inline void SetThreadName(Poco::Thread &thr, const char *name) {
-#ifdef __linux__
-		thr.setName(name);
-		pthread_setname_np(thr.tid(), name);
-#endif
-#ifdef __APPLE__
-		thr.setName(name);
-#endif
-	}
-
-    enum MediaTypeEncodings {
-        PLAIN,
-        BINARY,
-        BASE64
-    };
-    struct MediaTypeEncoding {
-        MediaTypeEncodings  Encoding=PLAIN;
-        std::string         ContentType;
-    };
-
-    [[nodiscard]] inline bool ValidSerialNumber(const std::string &Serial) {
-        return ((Serial.size() < uCentralProtocol::SERIAL_NUMBER_LENGTH) &&
-        std::all_of(Serial.begin(),Serial.end(),[](auto i){return std::isxdigit(i);}));
-    }
-
-    [[nodiscard]] inline bool ValidUUID(const std::string &UUID) {
-        if(UUID.size()>36)
-            return false;
-        uint dashes=0;
-        return (std::all_of(UUID.begin(),UUID.end(),[&](auto i){ if(i=='-') dashes++; return i=='-' || std::isxdigit(i);})) && (dashes>0);
-    }
-
-	template <typename ...Args> std::string ComputeHash(Args&&... args) {
-		Poco::SHA2Engine    E;
-		auto as_string = [](auto p) {
-			if constexpr(std::is_arithmetic_v<decltype(p)>) {
-				return std::to_string(p);
-			} else {
-				return p;
-			}
-		};
-		(E.update(as_string(args)),...);
-		return Poco::SHA2Engine::digestToHex(E.digest());
-	}
-
-    [[nodiscard]] inline std::vector<std::string> Split(const std::string &List, char Delimiter=',' ) {
-        std::vector<std::string> ReturnList;
-
-        unsigned long P=0;
-
-        while(P<List.size())
-        {
-            unsigned long P2 = List.find_first_of(Delimiter, P);
-            if(P2==std::string::npos) {
-                ReturnList.push_back(List.substr(P));
-                break;
-            }
-            else
-                ReturnList.push_back(List.substr(P,P2-P));
-            P=P2+1;
-        }
-        return ReturnList;
-    }
-
-
-    [[nodiscard]] inline std::string FormatIPv6(const std::string & I )
-    {
-        if(I.substr(0,8) == "[::ffff:")
-        {
-            unsigned long PClosingBracket = I.find_first_of(']');
-
-            std::string ip = I.substr(8, PClosingBracket-8);
-            std::string port = I.substr(PClosingBracket+1);
-            return ip + port;
-        }
-
-        return I;
-    }
-
-    inline void padTo(std::string& str, size_t num, char paddingChar = '\0') {
-        str.append(num - str.length() % num, paddingChar);
-    }
-
-    [[nodiscard]] inline std::string SerialToMAC(const std::string &Serial) {
-        std::string R = Serial;
-
-        if(R.size()<12)
-            padTo(R,12,'0');
-        else if (R.size()>12)
-            R = R.substr(0,12);
-
-        char buf[18];
-
-        buf[0] = R[0]; buf[1] = R[1] ; buf[2] = ':' ;
-        buf[3] = R[2] ; buf[4] = R[3]; buf[5] = ':' ;
-        buf[6] = R[4]; buf[7] = R[5] ; buf[8] = ':' ;
-        buf[9] = R[6] ; buf[10]= R[7]; buf[11] = ':';
-        buf[12] = R[8] ; buf[13]= R[9]; buf[14] = ':';
-        buf[15] = R[10] ; buf[16]= R[11];buf[17] = 0;
-
-        return buf;
-    }
-
-    inline uint64_t MACToInt(const std::string &MAC) {
-        uint64_t Result = 0 ;
-        for(const auto &c:MAC) {
-            if(c==':')
-                continue;
-            Result <<= 4;
-            if(c>='0' && c<='9') {
-                Result += (c - '0');
-            } else if (c>='a' && c<='f') {
-                Result += (c-'a'+10);
-            } else if (c>='A' && c<='F') {
-                Result += (c-'A'+10);
-            }
-        }
-        return Result;
-    }
-
-    [[nodiscard]] inline std::string ToHex(const std::vector<unsigned char> & B) {
-        std::string R;
-        R.reserve(B.size()*2);
-
-        static const char hex[] = "0123456789abcdef";
-
-        for(const auto &i:B)
-        {
-            R += (hex[ (i & 0xf0) >> 4]);
-            R += (hex[ (i & 0x0f) ]);
-        }
-
-        return R;
-    }
-
-    inline static const char kEncodeLookup[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-    inline static const char kPadCharacter = '=';
-
-    using byte = std::uint8_t;
-
-    [[nodiscard]] inline std::string base64encode(const byte *input, uint32_t size) {
-        std::string encoded;
-        encoded.reserve(((size / 3) + (size % 3 > 0)) * 4);
-
-        std::uint32_t temp,i,ee;
-        ee = (size/3);
-
-        for (i = 0; i < 3*ee; ++i) {
-            temp = input[i++] << 16;
-            temp += input[i++] << 8;
-            temp += input[i];
-            encoded.append(1, kEncodeLookup[(temp & 0x00FC0000) >> 18]);
-            encoded.append(1, kEncodeLookup[(temp & 0x0003F000) >> 12]);
-            encoded.append(1, kEncodeLookup[(temp & 0x00000FC0) >> 6]);
-            encoded.append(1, kEncodeLookup[(temp & 0x0000003F)]);
-        }
-
-        switch (size % 3) {
-            case 1:
-                temp = input[i] << 16;
-                encoded.append(1, kEncodeLookup[(temp & 0x00FC0000) >> 18]);
-                encoded.append(1, kEncodeLookup[(temp & 0x0003F000) >> 12]);
-                encoded.append(2, kPadCharacter);
-                break;
-                case 2:
-                    temp = input[i++] << 16;
-                    temp += input[i] << 8;
-                    encoded.append(1, kEncodeLookup[(temp & 0x00FC0000) >> 18]);
-                    encoded.append(1, kEncodeLookup[(temp & 0x0003F000) >> 12]);
-                    encoded.append(1, kEncodeLookup[(temp & 0x00000FC0) >> 6]);
-                    encoded.append(1, kPadCharacter);
-                    break;
-        }
-
-        return encoded;
-    }
-
-    [[nodiscard]] inline std::vector<byte> base64decode(const std::string& input)
-    {
-        if(input.length() % 4)
-            throw std::runtime_error("Invalid base64 length!");
-
-        std::size_t padding=0;
-
-        if(input.length())
-        {
-            if(input[input.length() - 1] == kPadCharacter) padding++;
-            if(input[input.length() - 2] == kPadCharacter) padding++;
-        }
-
-        std::vector<byte> decoded;
-        decoded.reserve(((input.length() / 4) * 3) - padding);
-
-        std::uint32_t temp=0;
-        auto it = input.begin();
-
-        while(it < input.end())
-        {
-            for(std::size_t i = 0; i < 4; ++i)
-            {
-                temp <<= 6;
-                if     (*it >= 0x41 && *it <= 0x5A) temp |= *it - 0x41;
-                else if(*it >= 0x61 && *it <= 0x7A) temp |= *it - 0x47;
-                else if(*it >= 0x30 && *it <= 0x39) temp |= *it + 0x04;
-                else if(*it == 0x2B)                temp |= 0x3E;
-                else if(*it == 0x2F)                temp |= 0x3F;
-                else if(*it == kPadCharacter)
-                {
-                    switch(input.end() - it)
-                    {
-                        case 1:
-                            decoded.push_back((temp >> 16) & 0x000000FF);
-                            decoded.push_back((temp >> 8 ) & 0x000000FF);
-                            return decoded;
-                            case 2:
-                                decoded.push_back((temp >> 10) & 0x000000FF);
-                                return decoded;
-                                default:
-                                    throw std::runtime_error("Invalid padding in base64!");
-                    }
-                }
-                else throw std::runtime_error("Invalid character in base64!");
-
-                ++it;
-            }
-
-            decoded.push_back((temp >> 16) & 0x000000FF);
-            decoded.push_back((temp >> 8 ) & 0x000000FF);
-            decoded.push_back((temp      ) & 0x000000FF);
-        }
-
-        return decoded;
-    }
-
-    inline bool ParseTime(const std::string &Time, int & Hours, int & Minutes, int & Seconds) {
-        Poco::StringTokenizer	TimeTokens(Time,":",Poco::StringTokenizer::TOK_TRIM);
-
-        Hours =  Minutes = Seconds = 0 ;
-        if(TimeTokens.count()==1) {
-            Hours 	= std::atoi(TimeTokens[0].c_str());
-        } else if(TimeTokens.count()==2) {
-            Hours 	= std::atoi(TimeTokens[0].c_str());
-            Minutes = std::atoi(TimeTokens[1].c_str());
-        } else if(TimeTokens.count()==3) {
-            Hours 	= std::atoi(TimeTokens[0].c_str());
-            Minutes = std::atoi(TimeTokens[1].c_str());
-            Seconds = std::atoi(TimeTokens[2].c_str());
-        } else
-            return false;
-        return true;
-    }
-
-
-    inline bool ParseDate(const std::string &Time, int & Year, int & Month, int & Day) {
-        Poco::StringTokenizer	DateTokens(Time,"-",Poco::StringTokenizer::TOK_TRIM);
-
-        Year =  Month = Day = 0 ;
-        if(DateTokens.count()==3) {
-            Year 	= std::atoi(DateTokens[0].c_str());
-            Month 	= std::atoi(DateTokens[1].c_str());
-            Day 	= std::atoi(DateTokens[2].c_str());
-        } else
-            return false;
-        return true;
-    }
-
-    inline bool CompareTime( int H1, int H2, int M1, int M2, int S1, int S2) {
-        if(H1<H2)
-            return true;
-        if(H1>H2)
-            return false;
-        if(M1<M2)
-            return true;
-        if(M2>M1)
-            return false;
-        if(S1<=S2)
-            return true;
-        return false;
-    }
-
-    [[nodiscard]] inline std::string LogLevelToString(int Level) {
-        switch(Level) {
-            case Poco::Message::PRIO_DEBUG: return "debug";
-            case Poco::Message::PRIO_INFORMATION: return "information";
-            case Poco::Message::PRIO_FATAL: return "fatal";
-            case Poco::Message::PRIO_WARNING: return "warning";
-            case Poco::Message::PRIO_NOTICE: return "notice";
-            case Poco::Message::PRIO_CRITICAL: return "critical";
-            case Poco::Message::PRIO_ERROR: return "error";
-            case Poco::Message::PRIO_TRACE: return "trace";
-            default: return "none";
-        }
-    }
-
-    [[nodiscard]] inline uint64_t SerialNumberToInt(const std::string & S) {
-		return std::stoull(S,nullptr,16);
-    }
-
-    [[nodiscard]] inline std::string IntToSerialNumber(uint64_t S) {
-        char b[16];
-        for(int i=0;i<12;++i) {
-            int B = (S & 0x0f);
-            if(B<10)
-                b[11-i] = B+'0';
-            else
-                b[11-i] = B - 10 + 'a';
-            S >>= 4 ;
-        }
-        b[12]=0;
-        return b;
-    }
-
-
-    [[nodiscard]] inline bool SerialNumberMatch(const std::string &S1, const std::string &S2, int Bits=2) {
-        auto S1_i = SerialNumberToInt(S1);
-        auto S2_i = SerialNumberToInt(S2);
-        return ((S1_i>>Bits)==(S2_i>>Bits));
-    }
-
-    [[nodiscard]] inline uint64_t SerialNumberToOUI(const std::string & S) {
-        uint64_t Result = 0 ;
-        int Digits=0;
-
-        for(const auto &i:S) {
-            if(std::isxdigit(i)) {
-                if(i>='0' && i<='9') {
-                    Result <<=4;
-                    Result += i-'0';
-                } else if(i>='A' && i<='F') {
-                    Result <<=4;
-                    Result += i-'A'+10;
-                } else if(i>='a' && i<='f') {
-                    Result <<=4;
-                    Result += i-'a'+10;
-                }
-                Digits++;
-                if(Digits==6)
-                    break;
-            }
-        }
-        return Result;
-    }
-
-    [[nodiscard]] inline uint64_t GetDefaultMacAsInt64() {
-        uint64_t Result=0;
-        auto IFaceList = Poco::Net::NetworkInterface::list();
-
-        for(const auto &iface:IFaceList) {
-            if(iface.isRunning() && !iface.isLoopback()) {
-                auto MAC = iface.macAddress();
-                for (auto const &i : MAC) {
-                    Result <<= 8;
-                    Result += (uint8_t)i;
-                }
-                if (Result != 0)
-                    break;
-            }
-        }
-        return Result;
-    }
-
-    [[nodiscard]] inline uint64_t InitializeSystemId() {
-        std::random_device	RDev;
-        std::srand(RDev());
-        std::chrono::high_resolution_clock	Clock;
-        auto Now = Clock.now().time_since_epoch().count();
-        auto S = (GetDefaultMacAsInt64() + std::rand() + Now)  ;
-		OpenWifi::AppServiceRegistry().Set("systemid",S);
-        return S;
-    }
-
-    [[nodiscard]] inline uint64_t GetSystemId();
-
-    [[nodiscard]] inline bool ValidEMailAddress(const std::string &email) {
-        // define a regular expression
-        static const std::regex pattern
-                ("[_a-z0-9-]+(\\.[_a-z0-9-]+)*(\\+[a-z0-9-]+)?@[a-z0-9-]+(\\.[a-z0-9-]+)*");
-        // try to match the string with the regular expression
-        return std::regex_match(email, pattern);
-    }
-
-    [[nodiscard]] inline std::string LoadFile( const Poco::File & F) {
-        std::string Result;
-        try {
-            std::ostringstream OS;
-            std::ifstream   IF(F.path());
-            Poco::StreamCopier::copyStream(IF, OS);
-            Result = OS.str();
-        } catch (...) {
-
-        }
-        return Result;
-    }
-
-    inline void ReplaceVariables( std::string & Content , const Types::StringPairVec & P) {
-        for(const auto &[Variable,Value]:P) {
-            Poco::replaceInPlace(Content,"${" + Variable + "}", Value);
-        }
-    }
-
-    [[nodiscard]] inline MediaTypeEncoding FindMediaType(const Poco::File &F) {
-        const auto E = Poco::Path(F.path()).getExtension();
-        if(E=="png")
-            return MediaTypeEncoding{   .Encoding = BINARY,
-                                        .ContentType = "image/png" };
-        if(E=="gif")
-            return MediaTypeEncoding{   .Encoding = BINARY,
-                                        .ContentType = "image/gif" };
-        if(E=="jpeg" || E=="jpg")
-            return MediaTypeEncoding{   .Encoding = BINARY,
-                                        .ContentType = "image/jpeg" };
-        if(E=="svg" || E=="svgz")
-            return MediaTypeEncoding{   .Encoding = PLAIN,
-                                        .ContentType = "image/svg+xml" };
-        if(E=="html")
-            return MediaTypeEncoding{   .Encoding = PLAIN,
-                                        .ContentType = "text/html" };
-        if(E=="css")
-            return MediaTypeEncoding{   .Encoding = PLAIN,
-                                        .ContentType = "text/css" };
-        if(E=="js")
-            return MediaTypeEncoding{   .Encoding = PLAIN,
-                                        .ContentType = "application/javascript" };
-        return MediaTypeEncoding{       .Encoding = BINARY,
-                                        .ContentType = "application/octet-stream" };
-    }
-
-    [[nodiscard]] inline std::string BinaryFileToHexString(const Poco::File &F) {
-        static const char hex[] = "0123456789abcdef";
-        std::string Result;
-        try {
-            std::ifstream IF(F.path());
-
-            int Count = 0;
-            while (IF.good()) {
-                if (Count)
-                    Result += ", ";
-                if ((Count % 32) == 0)
-                    Result += "\r\n";
-                Count++;
-                unsigned char C = IF.get();
-                Result += "0x";
-                Result += (char) (hex[(C & 0xf0) >> 4]);
-                Result += (char) (hex[(C & 0x0f)]);
-            }
-        } catch(...) {
-
-        }
-        return Result;
-    }
-
-    [[nodiscard]] inline std::string SecondsToNiceText(uint64_t Seconds) {
-        std::string Result;
-        int Days = Seconds / (24*60*60);
-        Seconds -= Days * (24*60*60);
-        int Hours= Seconds / (60*60);
-        Seconds -= Hours * (60*60);
-        int Minutes = Seconds / 60;
-        Seconds -= Minutes * 60;
-        Result = std::to_string(Days) +" days, " + std::to_string(Hours) + ":" + std::to_string(Minutes) + ":" + std::to_string(Seconds);
-        return Result;
-    }
-
-    [[nodiscard]] inline bool wgets(const std::string &URL, std::string &Response) {
-        try {
-            Poco::URI uri(URL);
-            Poco::Net::HTTPSClientSession session(uri.getHost(), uri.getPort());
-
-            // prepare path
-            std::string path(uri.getPathAndQuery());
-            if (path.empty()) {
-                path = "/";
-            }
-
-            // send request
-            Poco::Net::HTTPRequest req(Poco::Net::HTTPRequest::HTTP_GET, path, Poco::Net::HTTPMessage::HTTP_1_1);
-            session.sendRequest(req);
-
-            Poco::Net::HTTPResponse res;
-            std::istream &is = session.receiveResponse(res);
-            std::ostringstream os;
-
-            Poco::StreamCopier::copyStream(is,os);
-            Response = os.str();
-
-            return true;
-        } catch (...) {
-
-        }
-        return false;
-    }
-
-    template< typename T >
-    std::string int_to_hex( T i )
-    {
-        std::stringstream stream;
-        stream << std::setfill ('0') << std::setw(12)
-        << std::hex << i;
-        return stream.str();
-    }
-
-    inline bool ExtractBase64CompressedData(const std::string &CompressedData,
-                                            std::string &UnCompressedData, uint64_t compress_sz ) {
-        std::istringstream ifs(CompressedData);
-        Poco::Base64Decoder b64in(ifs);
-        std::ostringstream ofs;
-        Poco::StreamCopier::copyStream(b64in, ofs);
-
-        int factor = 20;
-        unsigned long MaxSize = compress_sz ? (unsigned long) (compress_sz + 5000) : (unsigned long) (ofs.str().size() * factor);
-        while(true) {
-            std::vector<uint8_t> UncompressedBuffer(MaxSize);
-            unsigned long FinalSize = MaxSize;
-            auto status = uncompress((uint8_t *)&UncompressedBuffer[0], &FinalSize,
-                                     (uint8_t *)ofs.str().c_str(), ofs.str().size());
-            if(status==Z_OK) {
-                UncompressedBuffer[FinalSize] = 0;
-                UnCompressedData = (char *)&UncompressedBuffer[0];
-                return true;
-            }
-            if(status==Z_BUF_ERROR) {
-                if(factor<300) {
-                    factor+=10;
-                    MaxSize = ofs.str().size() * factor;
-                    continue;
-                } else {
-                    return false;
-                }
-            }
-            return false;
-        }
-        return false;
-    }
-
-}
 
 namespace OpenWifi {
-
-	static const std::string uSERVICE_SECURITY{"owsec"};
-	static const std::string uSERVICE_GATEWAY{"owgw"};
-	static const std::string uSERVICE_FIRMWARE{ "owfms"};
-    static const std::string uSERVICE_TOPOLOGY{ "owtopo"};
-    static const std::string uSERVICE_PROVISIONING{ "owprov"};
-    static const std::string uSERVICE_OWLS{ "owls"};
-    static const std::string uSERVICE_SUBCRIBER{ "owsub"};
-    static const std::string uSERVICE_INSTALLER{ "owinst"};
-    static const std::string uSERVICE_ANALYTICS{ "owanalytics"};
-	static const std::string uSERVICE_OWRRM{ "owrrm"};
 
 	class ConfigurationEntry {
 	  public:
@@ -1378,327 +278,6 @@ namespace OpenWifi {
 		Poco::Logger		&Logger_;
 	};
 
-	class MyPrivateKeyPassphraseHandler : public Poco::Net::PrivateKeyPassphraseHandler {
-	public:
-	    explicit MyPrivateKeyPassphraseHandler(const std::string &Password, Poco::Logger & Logger):
-	    PrivateKeyPassphraseHandler(true),
-	    Password_(Password),
-        Logger_(Logger)
-        {}
-	    void onPrivateKeyRequested([[maybe_unused]] const void * pSender,std::string & privateKey) {
-	        Logger_.information("Returning key passphrase.");
-	        privateKey = Password_;
-	    };
-		inline Poco::Logger & Logger() { return Logger_; }
-	private:
-	    std::string Password_;
-	    Poco::Logger & Logger_;
-	};
-
-	class PropertiesFileServerEntry {
-	public:
-	    PropertiesFileServerEntry(std::string Address, uint32_t port, std::string Key_file,
-                                  std::string Cert_file, std::string RootCa, std::string Issuer,
-                                  std::string ClientCas, std::string Cas,
-                                  std::string Key_file_password = "", std::string Name = "",
-                                  Poco::Net::Context::VerificationMode M =
-                                          Poco::Net::Context::VerificationMode::VERIFY_RELAXED,
-                                          int backlog = 64)
-                                          : address_(std::move(Address)),
-                                          port_(port),
-                                          cert_file_(std::move(Cert_file)),
-                                          key_file_(std::move(Key_file)),
-                                          root_ca_(std::move(RootCa)),
-                                          key_file_password_(std::move(Key_file_password)),
-                                          issuer_cert_file_(std::move(Issuer)),
-                                          client_cas_(std::move(ClientCas)),
-                                          cas_(std::move(Cas)),
-                                          name_(std::move(Name)),
-                                          backlog_(backlog),
-                                          level_(M)
-                                          {};
-
-	    [[nodiscard]] inline const std::string &Address() const { return address_; };
-	    [[nodiscard]] inline uint32_t Port() const { return port_; };
-	    [[nodiscard]] inline auto KeyFile() const { return key_file_; };
-	    [[nodiscard]] inline auto CertFile() const { return cert_file_; };
-	    [[nodiscard]] inline auto RootCA() const { return root_ca_; };
-	    [[nodiscard]] inline auto KeyFilePassword() const { return key_file_password_; };
-	    [[nodiscard]] inline auto IssuerCertFile() const { return issuer_cert_file_; };
-	    [[nodiscard]] inline auto Name() const { return name_; };
-	    [[nodiscard]] inline int Backlog() const { return backlog_; }
-		[[nodiscard]] inline auto Cas() const { return cas_; }
-
-	    [[nodiscard]] inline Poco::Net::SecureServerSocket CreateSecureSocket(Poco::Logger &L) const {
-	        Poco::Net::Context::Params P;
-
-	        P.verificationMode = level_;
-	        P.verificationDepth = 9;
-	        P.loadDefaultCAs = root_ca_.empty();
-	        P.cipherList = "ALL:!ADH:!LOW:!EXP:!MD5:@STRENGTH";
-	        P.dhUse2048Bits = true;
-	        P.caLocation = cas_;
-
-	        auto Context = Poco::AutoPtr<Poco::Net::Context>(new Poco::Net::Context(Poco::Net::Context::TLS_SERVER_USE, P));
-
-	        if(!key_file_password_.empty()) {
-	            auto PassphraseHandler = Poco::SharedPtr<MyPrivateKeyPassphraseHandler>( new MyPrivateKeyPassphraseHandler(key_file_password_,L));
-	            Poco::Net::SSLManager::instance().initializeServer(PassphraseHandler, nullptr,Context);
-	        }
-
-	        if (!cert_file_.empty() && !key_file_.empty()) {
-	            Poco::Crypto::X509Certificate Cert(cert_file_);
-	            Poco::Crypto::X509Certificate Root(root_ca_);
-
-	            Context->useCertificate(Cert);
-	            Context->addChainCertificate(Root);
-
-	            Context->addCertificateAuthority(Root);
-
-	            if (level_ == Poco::Net::Context::VERIFY_STRICT) {
-	                if (issuer_cert_file_.empty()) {
-	                    L.fatal("In strict mode, you must supply ans issuer certificate");
-	                }
-	                if (client_cas_.empty()) {
-	                    L.fatal("In strict mode, client cas must be supplied");
-	                }
-	                Poco::Crypto::X509Certificate Issuing(issuer_cert_file_);
-	                Context->addChainCertificate(Issuing);
-	                Context->addCertificateAuthority(Issuing);
-	            }
-
-	            Poco::Crypto::RSAKey Key("", key_file_, key_file_password_);
-	            Context->usePrivateKey(Key);
-
-	            SSL_CTX *SSLCtx = Context->sslContext();
-	            if (!SSL_CTX_check_private_key(SSLCtx)) {
-	                L.fatal(fmt::format("Wrong Certificate({}) for Key({})", cert_file_, key_file_));
-	            }
-
-	            SSL_CTX_set_verify(SSLCtx, SSL_VERIFY_PEER, nullptr);
-
-	            if (level_ == Poco::Net::Context::VERIFY_STRICT) {
-	                SSL_CTX_set_client_CA_list(SSLCtx, SSL_load_client_CA_file(client_cas_.c_str()));
-	            }
-	            SSL_CTX_enable_ct(SSLCtx, SSL_CT_VALIDATION_STRICT);
-	            SSL_CTX_dane_enable(SSLCtx);
-
-	            Context->enableSessionCache();
-	            Context->setSessionCacheSize(0);
-	            Context->setSessionTimeout(60);
-	            Context->enableExtendedCertificateVerification(true);
-	            Context->disableStatelessSessionResumption();
-	        }
-
-	        if (address_ == "*") {
-	            Poco::Net::IPAddress Addr(Poco::Net::IPAddress::wildcard(
-	                    Poco::Net::Socket::supportsIPv6() ? Poco::Net::AddressFamily::IPv6
-	                    : Poco::Net::AddressFamily::IPv4));
-	            Poco::Net::SocketAddress SockAddr(Addr, port_);
-
-	            return Poco::Net::SecureServerSocket(SockAddr, backlog_, Context);
-	        } else {
-	            Poco::Net::IPAddress Addr(address_);
-	            Poco::Net::SocketAddress SockAddr(Addr, port_);
-
-	            return Poco::Net::SecureServerSocket(SockAddr, backlog_, Context);
-	        }
-	    }
-
-        [[nodiscard]] inline Poco::Net::ServerSocket CreateSocket([[maybe_unused]] Poco::Logger &L) const {
-            Poco::Net::Context::Params P;
-
-            if (address_ == "*") {
-                Poco::Net::IPAddress Addr(Poco::Net::IPAddress::wildcard(
-                        Poco::Net::Socket::supportsIPv6() ? Poco::Net::AddressFamily::IPv6
-                                                          : Poco::Net::AddressFamily::IPv4));
-                Poco::Net::SocketAddress SockAddr(Addr, port_);
-                return Poco::Net::ServerSocket(SockAddr, backlog_);
-            } else {
-                Poco::Net::IPAddress Addr(address_);
-                Poco::Net::SocketAddress SockAddr(Addr, port_);
-                return Poco::Net::ServerSocket(SockAddr, backlog_);
-            }
-        }
-
-	    inline void LogCertInfo(Poco::Logger &L, const Poco::Crypto::X509Certificate &C) const {
-	        L.information("=============================================================================================");
-	        L.information(fmt::format(">          Issuer: {}", C.issuerName()));
-	        L.information("---------------------------------------------------------------------------------------------");
-	        L.information(fmt::format(">     Common Name: {}",
-                                       C.issuerName(Poco::Crypto::X509Certificate::NID_COMMON_NAME)));
-	        L.information(fmt::format(">         Country: {}",
-                                       C.issuerName(Poco::Crypto::X509Certificate::NID_COUNTRY)));
-	        L.information(fmt::format(">        Locality: {}",
-                                       C.issuerName(Poco::Crypto::X509Certificate::NID_LOCALITY_NAME)));
-	        L.information(fmt::format(">      State/Prov: {}",
-                                       C.issuerName(Poco::Crypto::X509Certificate::NID_STATE_OR_PROVINCE)));
-	        L.information(fmt::format(">        Org name: {}",
-                                       C.issuerName(Poco::Crypto::X509Certificate::NID_ORGANIZATION_NAME)));
-	        L.information(
-	                fmt::format(">        Org unit: {}",
-                                 C.issuerName(Poco::Crypto::X509Certificate::NID_ORGANIZATION_UNIT_NAME)));
-	        L.information(
-	                fmt::format(">           Email: {}",
-                                 C.issuerName(Poco::Crypto::X509Certificate::NID_PKCS9_EMAIL_ADDRESS)));
-	        L.information(fmt::format(">         Serial#: {}",
-                                       C.issuerName(Poco::Crypto::X509Certificate::NID_SERIAL_NUMBER)));
-	        L.information("---------------------------------------------------------------------------------------------");
-	        L.information(fmt::format(">         Subject: {}", C.subjectName()));
-	        L.information("---------------------------------------------------------------------------------------------");
-	        L.information(fmt::format(">     Common Name: {}",
-                                       C.subjectName(Poco::Crypto::X509Certificate::NID_COMMON_NAME)));
-	        L.information(fmt::format(">         Country: {}",
-                                       C.subjectName(Poco::Crypto::X509Certificate::NID_COUNTRY)));
-	        L.information(fmt::format(">        Locality: {}",
-                                       C.subjectName(Poco::Crypto::X509Certificate::NID_LOCALITY_NAME)));
-	        L.information(
-	                fmt::format(">      State/Prov: {}",
-                                 C.subjectName(Poco::Crypto::X509Certificate::NID_STATE_OR_PROVINCE)));
-	        L.information(
-	                fmt::format(">        Org name: {}",
-                                 C.subjectName(Poco::Crypto::X509Certificate::NID_ORGANIZATION_NAME)));
-	        L.information(
-	                fmt::format(">        Org unit: {}",
-                                 C.subjectName(Poco::Crypto::X509Certificate::NID_ORGANIZATION_UNIT_NAME)));
-	        L.information(
-	                fmt::format(">           Email: {}",
-                                 C.subjectName(Poco::Crypto::X509Certificate::NID_PKCS9_EMAIL_ADDRESS)));
-	        L.information(fmt::format(">         Serial#: {}",
-                                       C.subjectName(Poco::Crypto::X509Certificate::NID_SERIAL_NUMBER)));
-	        L.information("---------------------------------------------------------------------------------------------");
-	        L.information(fmt::format(">  Signature Algo: {}", C.signatureAlgorithm()));
-	        auto From = Poco::DateTimeFormatter::format(C.validFrom(), Poco::DateTimeFormat::HTTP_FORMAT);
-	        L.information(fmt::format(">      Valid from: {}", From));
-	        auto Expires =
-	                Poco::DateTimeFormatter::format(C.expiresOn(), Poco::DateTimeFormat::HTTP_FORMAT);
-	        L.information(fmt::format(">      Expires on: {}", Expires));
-	        L.information(fmt::format(">         Version: {}", (int)C.version()));
-	        L.information(fmt::format(">        Serial #: {}", C.serialNumber()));
-	        L.information("=============================================================================================");
-	    }
-
-	    inline void LogCert(Poco::Logger &L) const {
-	        try {
-	            Poco::Crypto::X509Certificate C(cert_file_);
-	            L.information("=============================================================================================");
-	            L.information("=============================================================================================");
-	            L.information(fmt::format("Certificate Filename: {}", cert_file_));
-	            LogCertInfo(L, C);
-	            L.information("=============================================================================================");
-
-	            if (!issuer_cert_file_.empty()) {
-	                Poco::Crypto::X509Certificate C1(issuer_cert_file_);
-	                L.information("=============================================================================================");
-	                L.information("=============================================================================================");
-	                L.information(fmt::format("Issues Certificate Filename: {}", issuer_cert_file_));
-	                LogCertInfo(L, C1);
-	                L.information("=============================================================================================");
-	            }
-
-	            if (!client_cas_.empty()) {
-	                std::vector<Poco::Crypto::X509Certificate> Certs =
-	                        Poco::Net::X509Certificate::readPEM(client_cas_);
-
-	                L.information("=============================================================================================");
-	                L.information("=============================================================================================");
-	                L.information(fmt::format("Client CAs Filename: {}", client_cas_));
-	                L.information("=============================================================================================");
-	                auto i = 1;
-	                for (const auto &C3 : Certs) {
-	                    L.information(fmt::format(" Index: {}", i));
-	                    L.information("=============================================================================================");
-	                    LogCertInfo(L, C3);
-	                    i++;
-	                }
-	                L.information("=============================================================================================");
-	            }
-
-	        } catch (const Poco::Exception &E) {
-	            L.log(E);
-	        }
-	    }
-
-	    inline void LogCas(Poco::Logger &L) const {
-	        try {
-	            std::vector<Poco::Crypto::X509Certificate> Certs =
-	                    Poco::Net::X509Certificate::readPEM(root_ca_);
-
-	            L.information("=============================================================================================");
-	            L.information("=============================================================================================");
-	            L.information(fmt::format("CA Filename: {}", root_ca_));
-	            L.information("=============================================================================================");
-	            auto i = 1;
-	            for (const auto &C : Certs) {
-	                L.information(fmt::format(" Index: {}", i));
-	                L.information("=============================================================================================");
-	                LogCertInfo(L, C);
-	                i++;
-	            }
-	            L.information("=============================================================================================");
-	        } catch (const Poco::Exception &E) {
-	            L.log(E);
-	        }
-	    }
-
-	private:
-	    std::string address_;
-        uint32_t port_;
-	    std::string cert_file_;
-	    std::string key_file_;
-	    std::string root_ca_;
-	    std::string key_file_password_;
-	    std::string issuer_cert_file_;
-	    std::string client_cas_;
-	    std::string cas_;
-	    std::string name_;
-	    int backlog_;
-	    Poco::Net::Context::VerificationMode level_;
-	};
-
-	class SubSystemServer : public Poco::Util::Application::Subsystem {
-	public:
-	    SubSystemServer(const std::string & Name, const std::string &LoggingPrefix,
-                        const std::string & SubSystemConfigPrefix);
-
-	    inline void initialize(Poco::Util::Application &self) override;
-	    inline void uninitialize() override {
-	    }
-	    inline void reinitialize([[maybe_unused]] Poco::Util::Application &self) override {
-	        Logger_->L_.information("Reloading of this subsystem is not supported.");
-	    }
-	    inline void defineOptions([[maybe_unused]] Poco::Util::OptionSet &options) override {
-	    }
-	    inline const std::string & Name() const { return Name_; };
-	    inline const char * name() const override { return Name_.c_str(); }
-
-	    inline const PropertiesFileServerEntry & Host(uint64_t index) { return ConfigServersList_[index]; };
-	    inline uint64_t HostSize() const { return ConfigServersList_.size(); }
-		inline Poco::Logger & Logger() const { return Logger_->L_; }
-	    inline void SetLoggingLevel(const std::string & levelName) {
-			Logger_->L_.setLevel(Poco::Logger::parseLevel(levelName));
-		}
-	    inline int GetLoggingLevel() { return Logger_->L_.getLevel(); }
-
-	    virtual int Start() = 0;
-	    virtual void Stop() = 0;
-
-		struct LoggerWrapper {
-			Poco::Logger & L_;
-			LoggerWrapper(Poco::Logger &L) :
-				L_(L) {}
-		};
-
-	protected:
-	    std::recursive_mutex Mutex_;
-        std::vector<PropertiesFileServerEntry> ConfigServersList_;
-
-    private:
-        std::unique_ptr<LoggerWrapper>  Logger_;
-	    std::string 					Name_;
-        std::string         			LoggerPrefix_;
-	    std::string 					SubSystemConfigPrefix_;
-	};
 
 	class RESTAPI_GenericServer {
 	public:
@@ -2529,88 +1108,6 @@ namespace OpenWifi {
 	                }
 	            }
 
-	class OpenAPIRequestGet {
-	public:
-	    explicit OpenAPIRequestGet( const std::string & Type,
-                                    const std::string & EndPoint,
-                                    const Types::StringPairVec & QueryData,
-                                    uint64_t msTimeout):
-                                    Type_(Type),
-                                    EndPoint_(EndPoint),
-                                    QueryData_(QueryData),
-                                    msTimeout_(msTimeout) {};
-	    inline Poco::Net::HTTPServerResponse::HTTPStatus Do(Poco::JSON::Object::Ptr &ResponseObject, const std::string & BearerToken = "");
-	private:
-	    std::string 			Type_;
-	    std::string 			EndPoint_;
-	    Types::StringPairVec 	QueryData_;
-	    uint64_t 				msTimeout_;
-	};
-
-	class OpenAPIRequestPut {
-	public:
-	    explicit OpenAPIRequestPut( const std::string & Type,
-                                    const std::string & EndPoint,
-                                    const Types::StringPairVec & QueryData,
-                                    const Poco::JSON::Object & Body,
-                                    uint64_t msTimeout):
-                                    Type_(Type),
-                                    EndPoint_(EndPoint),
-                                    QueryData_(QueryData),
-                                    msTimeout_(msTimeout),
-                                    Body_(Body){};
-
-	    inline Poco::Net::HTTPServerResponse::HTTPStatus Do(Poco::JSON::Object::Ptr &ResponseObject, const std::string & BearerToken = "");
-
-	private:
-	    std::string 			Type_;
-	    std::string 			EndPoint_;
-	    Types::StringPairVec 	QueryData_;
-	    uint64_t 				msTimeout_;
-	    Poco::JSON::Object      Body_;
-	};
-
-	class OpenAPIRequestPost {
-	public:
-	    explicit OpenAPIRequestPost( const std::string & Type,
-                                     const std::string & EndPoint,
-                                     const Types::StringPairVec & QueryData,
-                                     const Poco::JSON::Object & Body,
-                                     uint64_t msTimeout):
-                                     Type_(Type),
-                                     EndPoint_(EndPoint),
-                                     QueryData_(QueryData),
-                                     msTimeout_(msTimeout),
-                                     Body_(Body){};
-	    inline Poco::Net::HTTPServerResponse::HTTPStatus Do(Poco::JSON::Object::Ptr &ResponseObject, const std::string & BearerToken = "");
-	private:
-	    std::string 			Type_;
-	    std::string 			EndPoint_;
-	    Types::StringPairVec 	QueryData_;
-	    uint64_t 				msTimeout_;
-	    Poco::JSON::Object      Body_;
-	};
-
-    class OpenAPIRequestDelete {
-    public:
-        explicit OpenAPIRequestDelete( const std::string & Type,
-                                     const std::string & EndPoint,
-                                     const Types::StringPairVec & QueryData,
-                                     uint64_t msTimeout):
-                Type_(Type),
-                EndPoint_(EndPoint),
-                QueryData_(QueryData),
-                msTimeout_(msTimeout){};
-        inline Poco::Net::HTTPServerResponse::HTTPStatus Do(const std::string & BearerToken = "");
-
-    private:
-        std::string 			Type_;
-        std::string 			EndPoint_;
-        Types::StringPairVec 	QueryData_;
-        uint64_t 				msTimeout_;
-        Poco::JSON::Object      Body_;
-    };
-
 	class KafkaMessage: public Poco::Notification {
 	  public:
 		KafkaMessage( const std::string &Topic, const std::string &Key, const std::string & Payload) :
@@ -2859,100 +1356,6 @@ namespace OpenWifi {
 	};
 
 	inline auto KafkaManager() { return KafkaManager::instance(); }
-
-	class AuthClient : public SubSystemServer {
-	public:
-	    explicit AuthClient() noexcept:
-	    SubSystemServer("Authentication", "AUTH-CLNT", "authentication")
-	    {
-	    }
-
-	    static auto instance() {
-	        static auto instance_ = new AuthClient;
-	        return instance_;
-	    }
-
-	    inline int Start() override {
-	        return 0;
-	    }
-
-	    inline void Stop() override {
-			poco_information(Logger(),"Stopping...");
-            std::lock_guard	G(Mutex_);
-	        Cache_.clear();
-			poco_information(Logger(),"Stopped...");
-	    }
-
-	    inline void RemovedCachedToken(const std::string &Token) {
-	        Cache_.remove(Token);
-	    }
-
-	    inline static bool IsTokenExpired(const SecurityObjects::WebToken &T) {
-	        return ((T.expires_in_+T.created_) < OpenWifi::Now());
-	    }
-
-	    inline bool RetrieveTokenInformation(const std::string & SessionToken,
-											 SecurityObjects::UserInfoAndPolicy & UInfo,
-											 std::uint64_t TID,
-											 bool & Expired, bool & Contacted, bool Sub=false) {
-	        try {
-	            Types::StringPairVec QueryData;
-	            QueryData.push_back(std::make_pair("token",SessionToken));
-	            OpenAPIRequestGet	Req(    uSERVICE_SECURITY,
-                                             Sub ? "/api/v1/validateSubToken" : "/api/v1/validateToken",
-                                             QueryData,
-                                             10000);
-	            Poco::JSON::Object::Ptr Response;
-
-                auto StatusCode = Req.Do(Response);
-                if(StatusCode==Poco::Net::HTTPServerResponse::HTTP_GATEWAY_TIMEOUT) {
-                    Contacted = false;
-                    return false;
-                }
-
-                Contacted = true;
-	            if(StatusCode==Poco::Net::HTTPServerResponse::HTTP_OK) {
-	                if(Response->has("tokenInfo") && Response->has("userInfo")) {
-	                    UInfo.from_json(Response);
-	                    if(IsTokenExpired(UInfo.webtoken)) {
-	                        Expired = true;
-	                        return false;
-	                    }
-	                    Expired = false;
-	                    Cache_.update(SessionToken, UInfo);
-	                    return true;
-	                } else {
-                        return false;
-                    }
-	            }
-	        } catch (...) {
-				poco_error(Logger(),fmt::format("Failed to retrieve token={} for TID={}", SessionToken, TID));
-	        }
-	        Expired = false;
-	        return false;
-	    }
-
-        inline bool IsAuthorized(const std::string &SessionToken, SecurityObjects::UserInfoAndPolicy & UInfo,
-								 std::uint64_t TID,
-								 bool & Expired, bool & Contacted, bool Sub = false) {
-	        auto User = Cache_.get(SessionToken);
-	        if(!User.isNull()) {
-	            if(IsTokenExpired(User->webtoken)) {
-	                Expired = true;
-	                return false;
-	            }
-	            Expired = false;
-                UInfo = *User;
-                return true;
-	        }
-	        return RetrieveTokenInformation(SessionToken, UInfo, TID, Expired, Contacted, Sub);
-	    }
-
-	private:
-	    Poco::ExpireLRUCache<std::string,OpenWifi::SecurityObjects::UserInfoAndPolicy>      Cache_{512,1200000 };
-	};
-
-	inline auto AuthClient() { return AuthClient::instance(); }
 
 	class ALBRequestHandler: public Poco::Net::HTTPRequestHandler
 	        /// Return a HTML document with the current date and time.
@@ -3235,19 +1638,7 @@ namespace OpenWifi {
         static inline std::atomic_uint64_t  NextTransactionId_ = 1;
 	};
 
-	struct MicroServiceMeta {
-		uint64_t 		Id=0;
-		std::string 	Type;
-		std::string 	PrivateEndPoint;
-		std::string 	PublicEndPoint;
-		std::string 	AccessKey;
-		std::string		Version;
-		uint64_t 		LastUpdate=0;
-	};
-
 	class SubSystemServer;
-	typedef std::map<std::string, MicroServiceMeta>	MicroServiceMetaMap;
-	typedef std::vector<MicroServiceMeta>			MicroServiceMetaVec;
 	typedef std::vector<SubSystemServer *>          SubSystemVec;
 
 	class MicroService : public Poco::Util::ServerApplication {
@@ -3310,8 +1701,8 @@ namespace OpenWifi {
 
         static inline void Exit(int Reason);
 		inline void BusMessageReceived(const std::string &Key, const std::string & Payload);
-		inline MicroServiceMetaVec GetServices(const std::string & Type);
-		inline MicroServiceMetaVec GetServices();
+		inline Types::MicroServiceMetaVec GetServices(const std::string & Type);
+		inline Types::MicroServiceMetaVec GetServices();
 		inline void LoadConfigurationFile();
 		inline void Reload();
 		inline void LoadMyConfig();
@@ -3389,7 +1780,7 @@ namespace OpenWifi {
 		std::string                 WWWAssetsDir_;
 		Poco::Crypto::CipherFactory & CipherFactory_ = Poco::Crypto::CipherFactory::defaultFactory();
 		Poco::Crypto::Cipher        * Cipher_ = nullptr;
-		MicroServiceMetaMap			Services_;
+		Types::MicroServiceMetaMap	Services_;
 		std::string 				MyHash_;
 		std::string 				MyPrivateEndPoint_;
 		std::string 				MyPublicEndPoint_;
@@ -3443,7 +1834,7 @@ namespace OpenWifi {
 	                            poco_debug(logger(),fmt::format("Service {} ID={} leaving system.",Object->get(KafkaTopics::ServiceEvents::Fields::PRIVATE).toString(),ID));
 	                        } else if (Event == KafkaTopics::ServiceEvents::EVENT_JOIN || Event == KafkaTopics::ServiceEvents::EVENT_KEEP_ALIVE) {
 								poco_debug(logger(),fmt::format("Service {} ID={} joining system.",Object->get(KafkaTopics::ServiceEvents::Fields::PRIVATE).toString(),ID));
-	                            Services_[PrivateEndPoint] = MicroServiceMeta{
+	                            Services_[PrivateEndPoint] = Types::MicroServiceMeta{
 	                                .Id = ID,
 	                                .Type = Poco::toLower(Object->get(KafkaTopics::ServiceEvents::Fields::TYPE).toString()),
 	                                .PrivateEndPoint = Object->get(KafkaTopics::ServiceEvents::Fields::PRIVATE).toString(),
@@ -3494,11 +1885,11 @@ namespace OpenWifi {
 	    }
 	}
 
-	inline MicroServiceMetaVec MicroService::GetServices(const std::string & Type) {
+	inline Types::MicroServiceMetaVec MicroService::GetServices(const std::string & Type) {
 	    std::lock_guard G(InfraMutex_);
 
 	    auto T = Poco::toLower(Type);
-	    MicroServiceMetaVec	Res;
+	    Types::MicroServiceMetaVec	Res;
 	    for(const auto &[_,ServiceRec]:Services_) {
 	        if(ServiceRec.Type==T)
 	            Res.push_back(ServiceRec);
@@ -3506,10 +1897,10 @@ namespace OpenWifi {
 	    return Res;
 	}
 
-	inline MicroServiceMetaVec MicroService::GetServices() {
+	inline Types::MicroServiceMetaVec MicroService::GetServices() {
 	    std::lock_guard G(InfraMutex_);
 
-	    MicroServiceMetaVec	Res;
+	    Types::MicroServiceMetaVec	Res;
 	    for(const auto &[_,ServiceRec]:Services_) {
 	        Res.push_back(ServiceRec);
 	    }
@@ -3961,13 +2352,6 @@ namespace OpenWifi {
 	    }
 	}
 
-    inline SubSystemServer::SubSystemServer(const std::string &Name, const std::string &LoggingPrefix,
-            const std::string &SubSystemConfigPrefix):
-		Name_(Name),
-		LoggerPrefix_(LoggingPrefix),
-		SubSystemConfigPrefix_(SubSystemConfigPrefix) {
-    }
-
     inline int RESTAPI_ExtServer::Start() {
 		Logger().information("Starting.");
         Server_.InitLogging();
@@ -4064,82 +2448,6 @@ namespace OpenWifi {
 	    }
 
 	    return Application::EXIT_OK;
-	}
-
-	AppServiceRegistry::AppServiceRegistry() {
-		FileName = MicroService::instance().DataDir() + "/registry.json";
-		Poco::File F(FileName);
-
-		try {
-			if(F.exists()) {
-				std::ostringstream  OS;
-				std::ifstream       IF(FileName);
-				Poco::StreamCopier::copyStream(IF, OS);
-				Registry_ = nlohmann::json::parse(OS.str());
-			}
-		} catch (...) {
-			Registry_ = nlohmann::json::parse("{}");
-		}
-	}
-
-	inline void SubSystemServer::initialize([[maybe_unused]] Poco::Util::Application &self) {
-	    auto i = 0;
-	    bool good = true;
-
-		auto NewLevel = MicroService::instance().ConfigGetString("logging.level." + Name_, "");
-		if(NewLevel.empty())
-        	Logger_ = std::make_unique<LoggerWrapper>(Poco::Logger::create(LoggerPrefix_, Poco::Logger::root().getChannel(), Poco::Logger::root().getLevel()));
-		else
-			Logger_ = std::make_unique<LoggerWrapper>(Poco::Logger::create(LoggerPrefix_, Poco::Logger::root().getChannel(), Poco::Logger::parseLevel(NewLevel)));
-
-	    ConfigServersList_.clear();
-	    while (good) {
-	        std::string root{SubSystemConfigPrefix_ + ".host." + std::to_string(i) + "."};
-
-	        std::string address{root + "address"};
-	        if (MicroService::instance().ConfigGetString(address, "").empty()) {
-	            good = false;
-	        } else {
-	            std::string port{root + "port"};
-	            std::string key{root + "key"};
-	            std::string key_password{root + "key.password"};
-	            std::string cert{root + "cert"};
-	            std::string name{root + "name"};
-	            std::string backlog{root + "backlog"};
-	            std::string rootca{root + "rootca"};
-	            std::string issuer{root + "issuer"};
-	            std::string clientcas(root + "clientcas");
-	            std::string cas{root + "cas"};
-
-	            std::string level{root + "security"};
-	            Poco::Net::Context::VerificationMode M = Poco::Net::Context::VERIFY_RELAXED;
-
-	            auto L = MicroService::instance().ConfigGetString(level, "");
-
-	            if (L == "strict") {
-	                M = Poco::Net::Context::VERIFY_STRICT;
-	            } else if (L == "none") {
-	                M = Poco::Net::Context::VERIFY_NONE;
-	            } else if (L == "relaxed") {
-	                M = Poco::Net::Context::VERIFY_RELAXED;
-	            } else if (L == "once")
-	                M = Poco::Net::Context::VERIFY_ONCE;
-
-	            PropertiesFileServerEntry entry(MicroService::instance().ConfigGetString(address, ""),
-                                                MicroService::instance().ConfigGetInt(port, 0),
-                                                MicroService::instance().ConfigPath(key, ""),
-                                                MicroService::instance().ConfigPath(cert, ""),
-                                                MicroService::instance().ConfigPath(rootca, ""),
-                                                MicroService::instance().ConfigPath(issuer, ""),
-                                                MicroService::instance().ConfigPath(clientcas, ""),
-                                                MicroService::instance().ConfigPath(cas, ""),
-                                                MicroService::instance().ConfigGetString(key_password, ""),
-                                                MicroService::instance().ConfigGetString(name, ""), M,
-                                                (int)MicroService::instance().ConfigGetInt(backlog, 64));
-	            ConfigServersList_.push_back(entry);
-	            i++;
-	        }
-	    }
 	}
 
 	inline int ALBHealthCheckServer::Start() {
@@ -4516,275 +2824,6 @@ namespace OpenWifi {
 	    void DoDelete() final {};
 	};
 
-    inline Poco::Net::HTTPServerResponse::HTTPStatus OpenAPIRequestGet::Do(Poco::JSON::Object::Ptr &ResponseObject, const std::string & BearerToken) {
-        try {
-            auto Services = MicroService::instance().GetServices(Type_);
-            for(auto const &Svc:Services) {
-                Poco::URI	URI(Svc.PrivateEndPoint);
-
-                auto Secure = (URI.getScheme() == "https");
-
-                URI.setPath(EndPoint_);
-                for (const auto &qp : QueryData_)
-                    URI.addQueryParameter(qp.first, qp.second);
-
-                std::string Path(URI.getPathAndQuery());
-                Poco::Net::HTTPRequest Request(Poco::Net::HTTPRequest::HTTP_GET,
-                                               Path,
-                                               Poco::Net::HTTPMessage::HTTP_1_1);
-
-                poco_debug(Poco::Logger::get("REST-CALLER-GET"),fmt::format(" {}", URI.toString()));
-
-                if(BearerToken.empty()) {
-                    Request.add("X-API-KEY", Svc.AccessKey);
-                    Request.add("X-INTERNAL-NAME", MicroService::instance().PublicEndPoint());
-                } else {
-                    // Authorization: Bearer ${token}
-                    Request.add("Authorization", "Bearer " + BearerToken);
-                }
-
-                if(Secure) {
-                    Poco::Net::HTTPSClientSession Session(URI.getHost(), URI.getPort());
-                    Session.setTimeout(Poco::Timespan(msTimeout_ / 1000, msTimeout_ % 1000));
-
-                    Session.sendRequest(Request);
-
-                    Poco::Net::HTTPResponse Response;
-                    std::istream &is = Session.receiveResponse(Response);
-                    if (Response.getStatus() == Poco::Net::HTTPResponse::HTTP_OK) {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    }
-                    return Response.getStatus();
-                } else {
-                    Poco::Net::HTTPClientSession Session(URI.getHost(), URI.getPort());
-                    Session.setTimeout(Poco::Timespan(msTimeout_ / 1000, msTimeout_ % 1000));
-
-                    Session.sendRequest(Request);
-
-                    Poco::Net::HTTPResponse Response;
-                    std::istream &is = Session.receiveResponse(Response);
-                    if (Response.getStatus() == Poco::Net::HTTPResponse::HTTP_OK) {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    }
-                    return Response.getStatus();
-                }
-            }
-        }
-        catch (const Poco::Exception &E)
-        {
-            Poco::Logger::get("REST-CALLER-GET").log(E);
-        }
-        return Poco::Net::HTTPServerResponse::HTTP_GATEWAY_TIMEOUT;
-    }
-
-    inline Poco::Net::HTTPServerResponse::HTTPStatus OpenAPIRequestPut::Do(Poco::JSON::Object::Ptr &ResponseObject, const std::string & BearerToken) {
-        try {
-            auto Services = MicroService::instance().GetServices(Type_);
-            for(auto const &Svc:Services) {
-                Poco::URI	URI(Svc.PrivateEndPoint);
-
-                auto Secure = (URI.getScheme() == "https");
-
-                URI.setPath(EndPoint_);
-                for (const auto &qp : QueryData_)
-                    URI.addQueryParameter(qp.first, qp.second);
-
-                poco_debug(Poco::Logger::get("REST-CALLER-PUT"),fmt::format("{}", URI.toString()));
-
-                std::string Path(URI.getPathAndQuery());
-
-                Poco::Net::HTTPRequest Request(Poco::Net::HTTPRequest::HTTP_PUT,
-                                               Path,
-                                               Poco::Net::HTTPMessage::HTTP_1_1);
-                std::ostringstream obody;
-                Poco::JSON::Stringifier::stringify(Body_,obody);
-
-                Request.setContentType("application/json");
-                Request.setContentLength(obody.str().size());
-
-                if(BearerToken.empty()) {
-                    Request.add("X-API-KEY", Svc.AccessKey);
-                    Request.add("X-INTERNAL-NAME", MicroService::instance().PublicEndPoint());
-                } else {
-                    // Authorization: Bearer ${token}
-                    Request.add("Authorization", "Bearer " + BearerToken);
-                }
-
-                if(Secure) {
-                    Poco::Net::HTTPSClientSession Session(URI.getHost(), URI.getPort());
-                    Session.setTimeout(Poco::Timespan(msTimeout_ / 1000, msTimeout_ % 1000));
-
-                    std::ostream &os = Session.sendRequest(Request);
-                    os << obody.str();
-
-                    Poco::Net::HTTPResponse Response;
-                    std::istream &is = Session.receiveResponse(Response);
-                    if (Response.getStatus() == Poco::Net::HTTPResponse::HTTP_OK) {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    } else {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    }
-                    return Response.getStatus();
-                } else {
-                    Poco::Net::HTTPClientSession Session(URI.getHost(), URI.getPort());
-                    Session.setTimeout(Poco::Timespan(msTimeout_ / 1000, msTimeout_ % 1000));
-
-                    std::ostream &os = Session.sendRequest(Request);
-                    os << obody.str();
-
-                    Poco::Net::HTTPResponse Response;
-                    std::istream &is = Session.receiveResponse(Response);
-                    if (Response.getStatus() == Poco::Net::HTTPResponse::HTTP_OK) {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    } else {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    }
-                    return Response.getStatus();
-                }
-            }
-        }
-        catch (const Poco::Exception &E)
-        {
-            Poco::Logger::get("REST-CALLER-PUT").log(E);
-        }
-        return Poco::Net::HTTPServerResponse::HTTP_GATEWAY_TIMEOUT;
-    }
-
-    inline Poco::Net::HTTPServerResponse::HTTPStatus OpenAPIRequestPost::Do(Poco::JSON::Object::Ptr &ResponseObject, const std::string & BearerToken) {
-        try {
-            auto Services = MicroService::instance().GetServices(Type_);
-
-            for(auto const &Svc:Services) {
-                Poco::URI	URI(Svc.PrivateEndPoint);
-
-
-                auto Secure = (URI.getScheme() == "https");
-
-                URI.setPath(EndPoint_);
-                for (const auto &qp : QueryData_)
-                    URI.addQueryParameter(qp.first, qp.second);
-
-                poco_debug(Poco::Logger::get("REST-CALLER-POST"),fmt::format(" {}", URI.toString()));
-
-                std::string Path(URI.getPathAndQuery());
-
-                Poco::Net::HTTPRequest Request(Poco::Net::HTTPRequest::HTTP_POST,
-                                               Path,
-                                               Poco::Net::HTTPMessage::HTTP_1_1);
-                std::ostringstream obody;
-                Poco::JSON::Stringifier::stringify(Body_,obody);
-
-                Request.setContentType("application/json");
-                Request.setContentLength(obody.str().size());
-
-                if(BearerToken.empty()) {
-                    Request.add("X-API-KEY", Svc.AccessKey);
-                    Request.add("X-INTERNAL-NAME", MicroService::instance().PublicEndPoint());
-                } else {
-                    // Authorization: Bearer ${token}
-                    Request.add("Authorization", "Bearer " + BearerToken);
-                }
-
-                if(Secure) {
-                    Poco::Net::HTTPSClientSession Session(URI.getHost(), URI.getPort());
-                    Session.setTimeout(Poco::Timespan(msTimeout_ / 1000, msTimeout_ % 1000));
-                    std::ostream &os = Session.sendRequest(Request);
-                    os << obody.str();
-
-                    Poco::Net::HTTPResponse Response;
-                    std::istream &is = Session.receiveResponse(Response);
-                    if (Response.getStatus() == Poco::Net::HTTPResponse::HTTP_OK) {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    } else {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    }
-                    return Response.getStatus();
-                } else {
-                    Poco::Net::HTTPClientSession Session(URI.getHost(), URI.getPort());
-                    Session.setTimeout(Poco::Timespan(msTimeout_ / 1000, msTimeout_ % 1000));
-                    std::ostream &os = Session.sendRequest(Request);
-                    os << obody.str();
-
-                    Poco::Net::HTTPResponse Response;
-                    std::istream &is = Session.receiveResponse(Response);
-                    if (Response.getStatus() == Poco::Net::HTTPResponse::HTTP_OK) {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    } else {
-                        Poco::JSON::Parser P;
-                        ResponseObject = P.parse(is).extract<Poco::JSON::Object::Ptr>();
-                    }
-                    return Response.getStatus();
-                }
-            }
-        }
-        catch (const Poco::Exception &E)
-        {
-            Poco::Logger::get("REST-CALLER-POST").log(E);
-        }
-        return Poco::Net::HTTPServerResponse::HTTP_GATEWAY_TIMEOUT;
-    }
-
-    inline Poco::Net::HTTPServerResponse::HTTPStatus OpenAPIRequestDelete::Do(const std::string & BearerToken) {
-        try {
-            auto Services = MicroService::instance().GetServices(Type_);
-
-            for(auto const &Svc:Services) {
-                Poco::URI	URI(Svc.PrivateEndPoint);
-
-                auto Secure = (URI.getScheme() == "https");
-
-                URI.setPath(EndPoint_);
-                for (const auto &qp : QueryData_)
-                    URI.addQueryParameter(qp.first, qp.second);
-
-                poco_debug(Poco::Logger::get("REST-CALLER-DELETE"),fmt::format(" {}", URI.toString()));
-
-                std::string Path(URI.getPathAndQuery());
-
-                Poco::Net::HTTPRequest Request(Poco::Net::HTTPRequest::HTTP_DELETE,
-                                               Path,
-                                               Poco::Net::HTTPMessage::HTTP_1_1);
-                if(BearerToken.empty()) {
-                    Request.add("X-API-KEY", Svc.AccessKey);
-                    Request.add("X-INTERNAL-NAME", MicroService::instance().PublicEndPoint());
-                } else {
-                    // Authorization: Bearer ${token}
-                    Request.add("Authorization", "Bearer " + BearerToken);
-                }
-
-                if(Secure) {
-                    Poco::Net::HTTPSClientSession Session(URI.getHost(), URI.getPort());
-                    Session.setTimeout(Poco::Timespan(msTimeout_ / 1000, msTimeout_ % 1000));
-                    Session.sendRequest(Request);
-                    Poco::Net::HTTPResponse Response;
-                    Session.receiveResponse(Response);
-                    return Response.getStatus();
-                } else {
-                    Poco::Net::HTTPClientSession Session(URI.getHost(), URI.getPort());
-                    Session.setTimeout(Poco::Timespan(msTimeout_ / 1000, msTimeout_ % 1000));
-                    Session.sendRequest(Request);
-                    Poco::Net::HTTPResponse Response;
-                    Session.receiveResponse(Response);
-                    return Response.getStatus();
-                }
-            }
-        }
-        catch (const Poco::Exception &E)
-        {
-            Poco::Logger::get("REST-CALLER-DELETE").log(E);
-        }
-        return Poco::Net::HTTPServerResponse::HTTP_GATEWAY_TIMEOUT;
-    }
-
     inline void RESTAPI_GenericServer::InitLogging() {
         std::string Public = MicroService::instance().ConfigGetString("apilogging.public.methods","PUT,POST,DELETE");
         SetFlags(true, Public);
@@ -4865,388 +2904,6 @@ namespace OpenWifi {
 
 	inline MicroService * MicroService::instance_ = nullptr;
 
-    template <typename ContentStruct> struct WebSocketNotification {
-        inline static uint64_t          xid=1;
-        uint64_t                        notification_id=++xid;
-		std::string						type;
-		ContentStruct    				content;
-
-        void to_json(Poco::JSON::Object &Obj) const;
-        bool from_json(const Poco::JSON::Object::Ptr &Obj);
-    };
-
-    template <typename ContentStruct> void WebSocketNotification<ContentStruct>::to_json(Poco::JSON::Object &Obj) const {
-        RESTAPI_utils::field_to_json(Obj,"notification_id",notification_id);
-		RESTAPI_utils::field_to_json(Obj,"type",type);
-		RESTAPI_utils::field_to_json(Obj,"content",content);
-    }
-
-	template <typename ContentStruct> bool WebSocketNotification<ContentStruct>::from_json(const Poco::JSON::Object::Ptr &Obj) {
-        try {
-            RESTAPI_utils::field_from_json(Obj,"notification_id",notification_id);
-            RESTAPI_utils::field_from_json(Obj,"content",content);
-			RESTAPI_utils::field_from_json(Obj,"type",type);
-            return true;
-        } catch(...) {
-
-        }
-        return false;
-    }
-
-    class WebSocketClientProcessor {
-    public:
-        virtual void Processor(const Poco::JSON::Object::Ptr &O, std::string &Answer, bool &Done ) = 0;
-    private:
-    };
-
-/*    class MyParallelSocketReactor {
-    public:
-        explicit MyParallelSocketReactor(uint32_t NumReactors = 8);
-        ~MyParallelSocketReactor();
-        Poco::Net::SocketReactor &Reactor();
-    private:
-        uint32_t NumReactors_;
-        Poco::Net::SocketReactor *Reactors_;
-        Poco::ThreadPool ReactorPool_;
-    };
-*/
-
-    class WebSocketClient;
-
-    class WebSocketClientServer : public SubSystemServer, Poco::Runnable {
-    public:
-        static auto instance() {
-            static auto instance_ = new WebSocketClientServer;
-            return instance_;
-        }
-
-        int Start() override;
-        void Stop() override;
-        void run() override;
-		Poco::Net::SocketReactor & Reactor() { return Reactor_; }
-        void NewClient(Poco::Net::WebSocket &WS, const std::string &Id, const std::string &UserName);
-        void SetProcessor(WebSocketClientProcessor *F);
-        void UnRegister(const std::string &Id);
-        void SetUser(const std::string &Id, const std::string &UserId);
-        [[nodiscard]] inline bool GeoCodeEnabled() const { return GeoCodeEnabled_; }
-        [[nodiscard]] inline std::string GoogleApiKey() const { return GoogleApiKey_; }
-        [[nodiscard]] bool Send(const std::string &Id, const std::string &Payload);
-
-		template <typename T> bool
-        	SendUserNotification(const std::string &userName, const WebSocketNotification<T> &Notification) {
-
-			Poco::JSON::Object  Payload;
-			Notification.to_json(Payload);
-			Poco::JSON::Object  Msg;
-			Msg.set("notification",Payload);
-			std::ostringstream OO;
-			Msg.stringify(OO);
-
-			return SendToUser(userName,OO.str());
-		}
-
-		template <typename T> void SendNotification(const WebSocketNotification<T> &Notification) {
-			Poco::JSON::Object  Payload;
-			Notification.to_json(Payload);
-			Poco::JSON::Object  Msg;
-			Msg.set("notification",Payload);
-			std::ostringstream OO;
-			Msg.stringify(OO);
-			SendToAll(OO.str());
-		}
-
-		[[nodiscard]] bool SendToUser(const std::string &userName, const std::string &Payload);
-		void SendToAll(const std::string &Payload);
-
-    private:
-        mutable std::atomic_bool Running_ = false;
-        Poco::Thread 								Thr_;
-		Poco::Net::SocketReactor					Reactor_;
-		Poco::Thread								ReactorThread_;
-        bool GeoCodeEnabled_ = false;
-        std::string GoogleApiKey_;
-        std::map<std::string, std::pair<std::unique_ptr<WebSocketClient>, std::string>> Clients_;
-        WebSocketClientProcessor *Processor_ = nullptr;
-        WebSocketClientServer() noexcept;
-    };
-
-    inline auto WebSocketClientServer() { return WebSocketClientServer::instance(); }
-
-    class WebSocketClient {
-    public:
-        explicit WebSocketClient(Poco::Net::WebSocket &WS,
-							   		const std::string &Id,
-							   		const std::string &UserName,
-							   		Poco::Logger &L,
-                                 	WebSocketClientProcessor *Processor);
-        virtual ~WebSocketClient();
-        [[nodiscard]] inline const std::string &Id();
-        [[nodiscard]] Poco::Logger &Logger();
-        inline bool Send(const std::string &Payload);
-		inline void EndConnection();
-    private:
-        std::unique_ptr<Poco::Net::WebSocket> WS_;
-        Poco::Net::SocketReactor 	&Reactor_;
-        std::string 				Id_;
-		std::string					UserName_;
-        Poco::Logger 				&Logger_;
-        std::atomic_bool 			Authenticated_ = false;
-		volatile bool				SocketRegistered_=false;
-        SecurityObjects::UserInfoAndPolicy UserInfo_;
-        WebSocketClientProcessor *Processor_ = nullptr;
-        void OnSocketReadable(const Poco::AutoPtr<Poco::Net::ReadableNotification> &pNf);
-        void OnSocketShutdown(const Poco::AutoPtr<Poco::Net::ShutdownNotification> &pNf);
-        void OnSocketError(const Poco::AutoPtr<Poco::Net::ErrorNotification> &pNf);
-    };
-
-    inline void WebSocketClientServer::NewClient(Poco::Net::WebSocket & WS, const std::string &Id, const std::string &UserName ) {
-        std::lock_guard G(Mutex_);
-        auto Client = std::make_unique<WebSocketClient>(WS,Id,UserName,Logger(), Processor_);
-        Clients_[Id] = std::make_pair(std::move(Client),"");
-    }
-
-    inline void WebSocketClientServer::SetProcessor( WebSocketClientProcessor * F) {
-        Processor_ = F;
-    }
-
-    inline void WebSocketClientServer::UnRegister(const std::string &Id) {
-        std::lock_guard G(Mutex_);
-        Clients_.erase(Id);
-    }
-
-    inline void WebSocketClientServer::SetUser(const std::string &Id, const std::string &UserId) {
-        std::lock_guard G(Mutex_);
-
-        auto it=Clients_.find(Id);
-        if(it!=Clients_.end()) {
-            Clients_[Id] = std::make_pair(std::move(it->second.first),UserId);
-        }
-    }
-
-    [[nodiscard]] inline bool SendToUser(const std::string &userName, const std::string &Payload);
-    inline WebSocketClientServer::WebSocketClientServer() noexcept:
-            SubSystemServer("WebSocketClientServer", "UI-WSCLNT-SVR", "websocketclients")
-    {
-    }
-
-    inline void WebSocketClientServer::run() {
-        Running_ = true ;
-		Utils::SetThreadName("ws:uiclnt-svr");
-        while(Running_) {
-            Poco::Thread::trySleep(2000);
-
-            if(!Running_)
-                break;
-        }
-    };
-
-    inline int WebSocketClientServer::Start() {
-        GoogleApiKey_ = MicroService::instance().ConfigGetString("google.apikey","");
-        GeoCodeEnabled_ = !GoogleApiKey_.empty();
-        // ReactorPool_ = std::make_unique<MyParallelSocketReactor>();
-		ReactorThread_.start(Reactor_);
-        Thr_.start(*this);
-        return 0;
-    };
-
-    inline void WebSocketClientServer::Stop() {
-        if(Running_) {
-			Reactor_.stop();
-			ReactorThread_.join();
-            Running_ = false;
-            Thr_.wakeUp();
-            Thr_.join();
-        }
-    };
-
-
-    inline void WebSocketClient::OnSocketError([[maybe_unused]] const Poco::AutoPtr<Poco::Net::ErrorNotification> &pNf) {
-        EndConnection();
-    }
-
-    inline bool WebSocketClientServer::Send(const std::string &Id, const std::string &Payload) {
-        std::lock_guard G(Mutex_);
-
-        auto It = Clients_.find(Id);
-        if(It!=Clients_.end())
-            return It->second.first->Send(Payload);
-        return false;
-    }
-
-    inline bool WebSocketClientServer::SendToUser(const std::string &UserName, const std::string &Payload) {
-        std::lock_guard G(Mutex_);
-        uint64_t Sent=0;
-
-        for(const auto &client:Clients_) {
-            if(client.second.second == UserName) {
-				try {
-					if (client.second.first->Send(Payload))
-						Sent++;
-				} catch (...) {
-					return false;
-				}
-            }
-        }
-        return Sent>0;
-    }
-
-	inline void WebSocketClientServer::SendToAll(const std::string &Payload) {
-		std::lock_guard G(Mutex_);
-
-		for(const auto &client:Clients_) {
-			try {
-				client.second.first->Send(Payload);
-			} catch (...) {
-
-			}
-		}
-	}
-
-	inline void WebSocketClient::OnSocketReadable([[maybe_unused]] const Poco::AutoPtr<Poco::Net::ReadableNotification> &pNf) {
-        int flags;
-        int n;
-        bool Done=false;
-		try {
-			Poco::Buffer<char> IncomingFrame(0);
-			n = WS_->receiveFrame(IncomingFrame, flags);
-			auto Op = flags & Poco::Net::WebSocket::FRAME_OP_BITMASK;
-
-			if (n == 0) {
-				poco_debug(Logger(),fmt::format("CLOSE({}): {} UI Client is closing WS connection.", Id_, UserName_));
-				return EndConnection();
-			}
-
-			switch (Op) {
-			case Poco::Net::WebSocket::FRAME_OP_PING: {
-				WS_->sendFrame("", 0,
-							   (int)Poco::Net::WebSocket::FRAME_OP_PONG |
-								   (int)Poco::Net::WebSocket::FRAME_FLAG_FIN);
-			} break;
-			case Poco::Net::WebSocket::FRAME_OP_PONG: {
-			} break;
-			case Poco::Net::WebSocket::FRAME_OP_CLOSE: {
-				poco_debug(Logger(),fmt::format("CLOSE({}): {} UI Client is closing WS connection.", Id_, UserName_));
-				Done = true;
-			} break;
-			case Poco::Net::WebSocket::FRAME_OP_TEXT: {
-				IncomingFrame.append(0);
-				if (!Authenticated_) {
-					std::string Frame{IncomingFrame.begin()};
-					auto Tokens = Utils::Split(Frame, ':');
-					bool Expired = false, Contacted = false;
-					if (Tokens.size() == 2 &&
-						AuthClient()->IsAuthorized(Tokens[1], UserInfo_, 0, Expired, Contacted)) {
-						Authenticated_ = true;
-						UserName_ = UserInfo_.userinfo.email;
-						poco_debug(Logger(),fmt::format("START({}): {} UI Client is starting WS connection.", Id_, UserName_));
-						std::string S{"Welcome! Bienvenue! Bienvenidos!"};
-						WS_->sendFrame(S.c_str(), S.size());
-						WebSocketClientServer()->SetUser(Id_, UserInfo_.userinfo.email);
-					} else {
-						std::string S{"Invalid token. Closing connection."};
-						WS_->sendFrame(S.c_str(), S.size());
-						Done = true;
-					}
-
-				} else {
-					try {
-						Poco::JSON::Parser P;
-						auto Obj =
-							P.parse(IncomingFrame.begin()).extract<Poco::JSON::Object::Ptr>();
-						std::string Answer;
-						if (Processor_ != nullptr)
-							Processor_->Processor(Obj, Answer, Done);
-						if (!Answer.empty())
-							WS_->sendFrame(Answer.c_str(), (int)Answer.size());
-						else {
-							WS_->sendFrame("{}", 2);
-						}
-					} catch (const Poco::JSON::JSONException &E) {
-						Logger().log(E);
-						Done=true;
-					}
-				}
-			} break;
-			default: {
-			}
-			}
-		} catch (...) {
-			Done=true;
-		}
-
-        if(Done) {
-            return EndConnection();
-        }
-    }
-
-    inline void WebSocketClient::OnSocketShutdown([[maybe_unused]] const Poco::AutoPtr<Poco::Net::ShutdownNotification> &pNf) {
-        EndConnection();
-    }
-
-
-    inline WebSocketClient::WebSocketClient( Poco::Net::WebSocket & WS , const std::string &Id, const std::string &UserName, Poco::Logger & L, WebSocketClientProcessor * Processor) :
-            Reactor_(WebSocketClientServer()->Reactor()),
-            Id_(Id),
-			UserName_(UserName),
-            Logger_(L),
-            Processor_(Processor) {
-		WS_ = std::make_unique<Poco::Net::WebSocket>(WS);
-		WS_->setNoDelay(true);
-		WS_->setKeepAlive(true);
-		WS_->setBlocking(false);
-		Reactor_.addEventHandler(*WS_,
-								 Poco::NObserver<WebSocketClient, Poco::Net::ReadableNotification>(
-										 *this, &WebSocketClient::OnSocketReadable));
-		Reactor_.addEventHandler(*WS_,
-								 Poco::NObserver<WebSocketClient, Poco::Net::ShutdownNotification>(
-										 *this, &WebSocketClient::OnSocketShutdown));
-		Reactor_.addEventHandler(*WS_,
-								 Poco::NObserver<WebSocketClient, Poco::Net::ErrorNotification>(
-										 *this, &WebSocketClient::OnSocketError));
-		SocketRegistered_ = true;
-		std::cout << "Starting client WS connection " << Id_ << std::endl;
-    }
-
-	inline void WebSocketClient::EndConnection() {
-		if(SocketRegistered_) {
-			SocketRegistered_ = false;
-			(*WS_).shutdown();
-			Reactor_.removeEventHandler(*WS_,
-										Poco::NObserver<WebSocketClient,
-														Poco::Net::ReadableNotification>(*this,&WebSocketClient::OnSocketReadable));
-			Reactor_.removeEventHandler(*WS_,
-										Poco::NObserver<WebSocketClient,
-														Poco::Net::ShutdownNotification>(*this,&WebSocketClient::OnSocketShutdown));
-			Reactor_.removeEventHandler(*WS_,
-										Poco::NObserver<WebSocketClient,
-														Poco::Net::ErrorNotification>(*this,&WebSocketClient::OnSocketError));
-			WebSocketClientServer()->UnRegister(Id_);
-		}
-	}
-
-    inline WebSocketClient::~WebSocketClient() {
-		EndConnection();
-		std::cout << "Deleting client WS connection " << Id_ << std::endl;
-    }
-
-    [[nodiscard]] inline const std::string & WebSocketClient::Id() {
-        return Id_;
-    };
-
-    [[nodiscard]] inline Poco::Logger & WebSocketClient::Logger() {
-        return Logger_;
-    }
-
-    [[nodiscard]] inline bool WebSocketClient::Send(const std::string &Payload) {
-        try {
-            WS_->sendFrame(Payload.c_str(),Payload.size());
-            return true;
-        } catch (...) {
-
-        }
-        return false;
-    }
-
     class RESTAPI_webSocketServer : public RESTAPIHandler {
     public:
         inline RESTAPI_webSocketServer(const RESTAPIHandler::BindingMap &bindings, Poco::Logger &L, RESTAPI_GenericServer &Server, uint64_t TransactionId, bool Internal)
@@ -5270,7 +2927,7 @@ namespace OpenWifi {
                 {
                     Poco::Net::WebSocket WS(*Request, *Response);
                     auto Id = MicroService::CreateUUID();
-                    WebSocketClientServer()->NewClient(WS,Id,UserInfo_.userinfo.email);
+                    UI_WebSocketClientServer()->NewClient(WS,Id,UserInfo_.userinfo.email);
                 }
                 catch (...) {
                     std::cout << "Cannot create websocket client..." << std::endl;
@@ -5282,16 +2939,6 @@ namespace OpenWifi {
     }
 
 
-}
-
-namespace OpenWifi::Utils {
-	[[nodiscard]] inline uint64_t GetSystemId() {
-		uint64_t ID=0;
-		if(!AppServiceRegistry().Get("systemid",ID)) {
-			return InitializeSystemId();
-		}
-		return ID;
-	}
 }
 
 namespace OpenWifi::CIDR {
