+++ src/framework/MicroServiceNames.h
@@ -1,0 +1,22 @@
+//
+// Created by stephane bourque on 2022-10-25.
+//
+
+#pragma once
+
+#include <string>
+
+namespace OpenWifi {
+
+	static const std::string uSERVICE_SECURITY{"owsec"};
+	static const std::string uSERVICE_GATEWAY{"owgw"};
+	static const std::string uSERVICE_FIRMWARE{ "owfms"};
+	static const std::string uSERVICE_TOPOLOGY{ "owtopo"};
+	static const std::string uSERVICE_PROVISIONING{ "owprov"};
+	static const std::string uSERVICE_OWLS{ "owls"};
+	static const std::string uSERVICE_SUBCRIBER{ "owsub"};
+	static const std::string uSERVICE_INSTALLER{ "owinst"};
+	static const std::string uSERVICE_ANALYTICS{ "owanalytics"};
+	static const std::string uSERVICE_OWRRM{ "owrrm"};
+
+}
