+++ src/framework/WebSocketClientNotifications.h
@@ -4,9 +4,38 @@
 
 #pragma once
 
-#include "framework/MicroService.h"
+#include "framework/RESTAPI_utils.h"
+#include "framework/utils.h"
 
 namespace OpenWifi {
+
+	template <typename ContentStruct> struct WebSocketNotification {
+		inline static uint64_t          xid=1;
+		uint64_t                        notification_id=++xid;
+		std::string						type;
+		ContentStruct    				content;
+
+		void to_json(Poco::JSON::Object &Obj) const;
+		bool from_json(const Poco::JSON::Object::Ptr &Obj);
+	};
+
+	template <typename ContentStruct> void WebSocketNotification<ContentStruct>::to_json(Poco::JSON::Object &Obj) const {
+		RESTAPI_utils::field_to_json(Obj,"notification_id",notification_id);
+		RESTAPI_utils::field_to_json(Obj,"type",type);
+		RESTAPI_utils::field_to_json(Obj,"content",content);
+	}
+
+	template <typename ContentStruct> bool WebSocketNotification<ContentStruct>::from_json(const Poco::JSON::Object::Ptr &Obj) {
+		try {
+			RESTAPI_utils::field_from_json(Obj,"notification_id",notification_id);
+			RESTAPI_utils::field_from_json(Obj,"content",content);
+			RESTAPI_utils::field_from_json(Obj,"type",type);
+			return true;
+		} catch(...) {
+
+		}
+		return false;
+	}
 
 	struct WebNotificationSingleDevice {
 		std::string		serialNumber;
@@ -93,47 +122,13 @@ namespace OpenWifi {
 		}
 	};
 
-	inline void  WebSocketClientNotificationNumberOfConnections(std::uint64_t numberOfDevices,
-															   std::uint64_t averageConnectedTime,
-															   std::uint64_t numberOfConnectingDevices) {
-		WebSocketNotification<WebSocketClientNotificationNumberOfConnection>	N;
-		N.content.numberOfDevices = numberOfDevices;
-		N.content.averageConnectedTime = averageConnectedTime;
-		N.content.numberOfConnectingDevices = numberOfConnectingDevices;
-		N.type = "device_connections_statistics";
-		WebSocketClientServer()->SendNotification(N);
-	}
-
-	inline void WebSocketClientNotificationDeviceConfigurationChange(const std::string &SerialNumber, uint64_t oldUUID, uint64_t newUUID) {
-		WebSocketNotification<WebNotificationSingleDeviceConfigurationChange>	N;
-		N.content.serialNumber = SerialNumber;
-		N.content.oldUUID = oldUUID;
-		N.content.newUUID = newUUID;
-		N.type = "device_configuration_upgrade";
-		WebSocketClientServer()->SendNotification(N);
-	}
-
-	inline void WebSocketClientNotificationDeviceFirmwareUpdated(const std::string &SerialNumber, const std::string &Firmware) {
-		WebSocketNotification<WebNotificationSingleDeviceFirmwareChange>	N;
-		N.content.serialNumber = SerialNumber;
-		N.content.newFirmware = Firmware;
-		N.type = "device_firmware_upgrade";
-		WebSocketClientServer()->SendNotification(N);
-	}
-
-	inline void WebSocketClientNotificationDeviceConnected(const std::string &SerialNumber) {
-		WebSocketNotification<WebNotificationSingleDevice>	N;
-		N.content.serialNumber = SerialNumber;
-		N.type = "device_connection";
-		WebSocketClientServer()->SendNotification(N);
-	}
-
-	inline void WebSocketClientNotificationDeviceDisconnected(const std::string & SerialNumber) {
-		WebSocketNotification<WebNotificationSingleDevice>	N;
-		N.content.serialNumber = SerialNumber;
-		N.type = "device_disconnection";
-		WebSocketClientServer()->SendNotification(N);
-	}
+	void WebSocketClientNotificationNumberOfConnections(std::uint64_t numberOfDevices,
+													   std::uint64_t averageConnectedTime,
+													   std::uint64_t numberOfConnectingDevices);
+	void WebSocketClientNotificationDeviceConfigurationChange(const std::string &SerialNumber, uint64_t oldUUID, uint64_t newUUID);
+	void WebSocketClientNotificationDeviceFirmwareUpdated(const std::string &SerialNumber, const std::string &Firmware);
+	void WebSocketClientNotificationDeviceConnected(const std::string &SerialNumber);
+	void WebSocketClientNotificationDeviceDisconnected(const std::string & SerialNumber);
 
     struct WebSocketNotificationJobContent {
         std::string                 title,
@@ -142,44 +137,15 @@ namespace OpenWifi {
         std::vector<std::string>    success,
                                     error,
                                     warning;
-        uint64_t                    timeStamp=OpenWifi::Now();
+        uint64_t                    timeStamp=OpenWifi::Utils::Now();
 
         void to_json(Poco::JSON::Object &Obj) const;
         bool from_json(const Poco::JSON::Object::Ptr &Obj);
     };
 
-    inline void WebSocketNotificationJobContent::to_json(Poco::JSON::Object &Obj) const {
-        RESTAPI_utils::field_to_json(Obj,"title",title);
-        RESTAPI_utils::field_to_json(Obj,"jobId",jobId);
-        RESTAPI_utils::field_to_json(Obj,"success",success);
-        RESTAPI_utils::field_to_json(Obj,"error",error);
-        RESTAPI_utils::field_to_json(Obj,"warning",warning);
-        RESTAPI_utils::field_to_json(Obj,"timeStamp",timeStamp);
-        RESTAPI_utils::field_to_json(Obj,"details",details);
-    }
-
-    inline bool WebSocketNotificationJobContent::from_json(const Poco::JSON::Object::Ptr &Obj) {
-        try {
-            RESTAPI_utils::field_from_json(Obj,"title",title);
-            RESTAPI_utils::field_from_json(Obj,"jobId",jobId);
-            RESTAPI_utils::field_from_json(Obj,"success",success);
-            RESTAPI_utils::field_from_json(Obj,"error",error);
-            RESTAPI_utils::field_from_json(Obj,"warning",warning);
-            RESTAPI_utils::field_from_json(Obj,"timeStamp",timeStamp);
-            RESTAPI_utils::field_from_json(Obj,"details",details);
-            return true;
-        } catch(...) {
-
-        }
-        return false;
-    }
-
     typedef WebSocketNotification<WebSocketNotificationJobContent>  WebSocketClientNotificationVenueUpdateJob_t;
 
-    inline void WebSocketClientNotificationVenueUpdateJobCompletionToUser( const std::string & User, WebSocketClientNotificationVenueUpdateJob_t &N) {
-        N.type = "venue_configuration_update";
-        WebSocketClientServer()->SendUserNotification(User,N);
-    }
+    void WebSocketClientNotificationVenueUpdateJobCompletionToUser( const std::string & User, WebSocketClientNotificationVenueUpdateJob_t &N);
 
     /////
     /////
@@ -191,7 +157,7 @@ namespace OpenWifi {
                 jobId;
         std::vector<std::string>    success,
                 warning;
-        uint64_t                    timeStamp=OpenWifi::Now();
+        uint64_t                    timeStamp=OpenWifi::Utils::Now();
 
         void to_json(Poco::JSON::Object &Obj) const;
         bool from_json(const Poco::JSON::Object::Ptr &Obj);
@@ -199,38 +165,7 @@ namespace OpenWifi {
 
     typedef WebSocketNotification<WebSocketNotificationRebootList> WebSocketClientNotificationVenueRebootList_t;
 
-    inline void WebSocketNotificationRebootList::to_json(Poco::JSON::Object &Obj) const {
-        RESTAPI_utils::field_to_json(Obj,"title",title);
-        RESTAPI_utils::field_to_json(Obj,"jobId",jobId);
-        RESTAPI_utils::field_to_json(Obj,"success",success);
-        RESTAPI_utils::field_to_json(Obj,"warning",warning);
-        RESTAPI_utils::field_to_json(Obj,"timeStamp",timeStamp);
-        RESTAPI_utils::field_to_json(Obj,"details",details);
-    }
-
-    inline bool WebSocketNotificationRebootList::from_json(const Poco::JSON::Object::Ptr &Obj) {
-        try {
-            RESTAPI_utils::field_from_json(Obj,"title",title);
-            RESTAPI_utils::field_from_json(Obj,"jobId",jobId);
-            RESTAPI_utils::field_from_json(Obj,"success",success);
-            RESTAPI_utils::field_from_json(Obj,"warning",warning);
-            RESTAPI_utils::field_from_json(Obj,"timeStamp",timeStamp);
-            RESTAPI_utils::field_from_json(Obj,"details",details);
-            return true;
-        } catch(...) {
-
-        }
-        return false;
-    }
-
-    inline void WebSocketClientNotificationVenueRebootCompletionToUser( const std::string & User, WebSocketClientNotificationVenueRebootList_t &N) {
-        N.type = "venue_rebooter";
-        WebSocketClientServer()->SendUserNotification(User,N);
-    }
-
-    /////
-    /////
-    /////
+    void WebSocketClientNotificationVenueRebootCompletionToUser( const std::string & User, WebSocketClientNotificationVenueRebootList_t &N);
 
     struct WebSocketNotificationUpgradeList {
         std::string                 title,
@@ -240,7 +175,7 @@ namespace OpenWifi {
                                     skipped,
                                     no_firmware,
                                     not_connected;
-        uint64_t                    timeStamp=OpenWifi::Now();
+        uint64_t                    timeStamp=OpenWifi::Utils::Now();
 
         void to_json(Poco::JSON::Object &Obj) const;
         bool from_json(const Poco::JSON::Object::Ptr &Obj);
@@ -248,38 +183,8 @@ namespace OpenWifi {
 
     typedef WebSocketNotification<WebSocketNotificationUpgradeList> WebSocketClientNotificationVenueUpgradeList_t;
 
-    inline void WebSocketNotificationUpgradeList::to_json(Poco::JSON::Object &Obj) const {
-        RESTAPI_utils::field_to_json(Obj,"title",title);
-        RESTAPI_utils::field_to_json(Obj,"jobId",jobId);
-        RESTAPI_utils::field_to_json(Obj,"success",success);
-        RESTAPI_utils::field_to_json(Obj,"notConnected",not_connected);
-        RESTAPI_utils::field_to_json(Obj,"noFirmware",no_firmware);
-        RESTAPI_utils::field_to_json(Obj,"skipped",skipped);
-        RESTAPI_utils::field_to_json(Obj,"timeStamp",timeStamp);
-        RESTAPI_utils::field_to_json(Obj,"details",details);
-    }
 
-    inline bool WebSocketNotificationUpgradeList::from_json(const Poco::JSON::Object::Ptr &Obj) {
-        try {
-            RESTAPI_utils::field_from_json(Obj,"title",title);
-            RESTAPI_utils::field_from_json(Obj,"jobId",jobId);
-            RESTAPI_utils::field_from_json(Obj,"success",success);
-            RESTAPI_utils::field_from_json(Obj,"notConnected",not_connected);
-            RESTAPI_utils::field_from_json(Obj,"noFirmware",no_firmware);
-            RESTAPI_utils::field_from_json(Obj,"skipped",skipped);
-            RESTAPI_utils::field_from_json(Obj,"timeStamp",timeStamp);
-            RESTAPI_utils::field_from_json(Obj,"details",details);
-            return true;
-        } catch(...) {
-
-        }
-        return false;
-    }
-
-    inline void WebSocketClientNotificationVenueUpgradeCompletionToUser( const std::string & User, WebSocketClientNotificationVenueUpgradeList_t &N) {
-        N.type = "venue_upgrader";
-        WebSocketClientServer()->SendUserNotification(User,N);
-    }
+    void WebSocketClientNotificationVenueUpgradeCompletionToUser( const std::string & User, WebSocketClientNotificationVenueUpgradeList_t &N);
 
 } // namespace OpenWifi
